msboot.relsurv.boot {mstate}	R Documentation

Default theta function used for msboot.relsurv

Description

Helper function used for calling inside msboot.relsurv (used for every bootstrap dataset). This function is used for calculating split hazards and evaluating them at all needed times.

Usage

msboot.relsurv.boot(
  data,
  transmat,
  all_times,
  split.transitions,
  rmap,
  time.format,
  boot_orig_msfit = FALSE,
  ratetable = relsurv::slopop,
  add.times
)

Arguments

data
An object of class 'msdata' containing a bootstrapped sample

transmat
The transition matrix of class transMat

all_times
All times at which the hazards have to be evaluated

split.transitions
An integer vector containing the numbered transitions that should be split. Use same numbering as in the given transition matrix

rmap
An optional list to be used if the variables in the dataset are not organized (and named) in the same way as in the ratetable object

time.format
Define the time format which is used in the dataset Possible options: c('days', 'years', 'months'). Default is 'days'

boot_orig_msfit
Logical, if true, do the bootstrap for the basic msfit model

ratetable
The population mortality table. A table of event rates, organized as a ratetable object, see for example relsurv::slopop. Default is slopop

add.times
Additional times at which hazards should be evaluated

Value

A list of calculated values for the given bootstrap sample.

Author(s)

Damjan Manevski damjan.manevski@mf.uni-lj.si

See Also

msboot.relsurv

[Package mstate version 0.3.3 Index]