print.summary.msfit {mstate}	R Documentation

Print method for summary.msfit object

Description

Print method for summary.msfit object

Usage

## S3 method for class 'summary.msfit'
print(x, complete = FALSE, ...)

Arguments

x
Object of class 'summary.msfit', to be printed

complete
Whether or not the complete estimated cumulative transition intensities should be printed (TRUE) or not (FALSE); default is FALSE, in which case the estimated cumulative transition hazards will be printed for the first and last 6 time points of each transition or of the selected times (or all when there are at most 12 of these time points

...
Further arguments to print

Examples

## Not run: 
# If all time points should be printed, specify complete=TRUE in the print statement
print(x, complete=TRUE)

## End(Not run)

[Package mstate version 0.3.3 Index]