varHaz.fixed {mstate}	R Documentation

Upgrade the varHaz object

Description

A function that upgrades varHaz from the msfit object where the variances are estimated using the Greenwood estimator; it is further assumed that variances for the population hazards are equal to zero.

Usage

varHaz.fixed(varHaz, link_trans, varHaz_original)

Arguments

varHaz
The varHaz object (present in a msfit object).

link_trans
A list that gives the linkage between the original and upgraded transition matrix.

varHaz_original
The original varHaz object from msfit (without the eventual time conversion).

Value

Return the upgraded varHaz object containing variances for the split transitions.

Author(s)

Damjan Manevski damjan.manevski@mf.uni-lj.si

[Package mstate version 0.3.3 Index]