Calculate nonparametric cumulative incidence functions and associated standard errors

This function computes nonparametric cumulative incidence functions and associated standard errors for each value of a group variable.

Cuminc(
  time,
  status,
  data,
  group,
  failcodes,
  na.status = c("remove", "extra"),
  variance = TRUE
)

Arguments

time
Either 1) a numeric vector containing the failure times or 2) a string containing the column name indicating these failure times

status
Either 1) a numeric, factor or character vector containing the failure codes or 2) a string containing the column name indicating these failure codes

data
When appropriate, a data frame containing time, status and/or group variables

group
Optionally, name of column in data indicating a grouping variable; cumulative incidence functions are calculated for each value or level of group. If missing no groups are considered

failcodes
A vector indicating which values of status are considered as different causes of failure; other values of status are considered as censorings. If missing and status is numeric, it is assumed that 0 is censoring and all other values indicate failcodes; if missing and status is character or factor, then it is assumed that each of the levels/values of status is a cause of failure

na.status
One of "remove" (default) or "extra", indicating whether subjects with missing cause of failure should be removed or whether missing cause of failure should be treated as a separate cause of failure

variance
Logical value, indicating whether the standard errors of the cumulative incidences should be output (TRUE, the default) or not

Value

An object of class "Cuminc", which is a data frame containing the estimated failure-free probabilities and cumulative incidences and their standard errors. The names of the dataframe are time, Surv, seSurv, and cuminc and secuminc followed by the values or levels of the failcodes. If group was specified, a group variable is included with the same name and values/levels as the original grouping variable, and with estimated cumulative incidences (SE) for each value/level of group.

Cuminc is now simply a wrapper around survfit of the survival package with type="mstate", only maintained for backward compatibility. The survfit object is kept as attribute (attr("survfit")), and the print, plot and summary functions are simply print, plot and summary applied to the survfit object. Subsetting the "Cuminc" object results in subsetting the data frame, not in subsetting the survfit object.

Details

The estimated cumulative incidences are as described in Putter, Fiocco & Geskus (2007); the standard errors are the square roots of the Greenwood variance estimators, see eg. Andersen, Borgan, Gill & Keiding (1993), de Wreede, Fiocco & Putter (2009), and they correspond to the variances in eg. Marubini & Valsecchi (1995). In case of no censoring, the estimated cumulative incidences and variances reduce to simple binomial frequencies and their variances.

References

Andersen PK, Borgan O, Gill RD, Keiding N (1993). Statistical Models Based on Counting Processes. Springer, New York.

Marubini E, Valsecchi MG (1995). Analysing Survival Data from Clinical Trials and Observational Studies. Wiley, New York.

Putter H, Fiocco M, Geskus RB (2007). Tutorial in biostatistics: Competing risks and multi-state models. Statistics in Medicine 26, 2389--2430.

de Wreede L, Fiocco M, Putter H (2009). The mstate package for estimation and prediction in non- and semi-parametric multi-state models. Submitted. http://www.msbi.nl/multistate.

Author

Hein Putter H.Putter@lumc.nl

Examples

### These data were used in Putter, Fiocco & Geskus (2007)
data(aidssi)
ci <- Cuminc(time=aidssi$time, status=aidssi$status)
head(ci); tail(ci)

#>    time      Surv CI.1        CI.2      seSurv seCI.1      seCI.2
#> 1 0.112 0.9969605    0 0.003039514 0.003034891      0 0.003034891
#> 2 0.137 0.9939210    0 0.006079027 0.004285436      0 0.004285436
#> 3 0.474 0.9908628    0 0.009137246 0.005251290      0 0.005251290
#> 4 0.824 0.9877760    0 0.012224046 0.006074796      0 0.006074796
#> 5 0.884 0.9846795    0 0.015320522 0.006799283      0 0.006799283
#> 6 0.969 0.9815830    0 0.018416998 0.007449696      0 0.007449696

#>       time      Surv      CI.1      CI.2     seSurv     seCI.1     seCI.2
#> 211 11.943 0.2312339 0.4035707 0.3651954 0.02638091 0.02978948 0.02881464
#> 212 12.129 0.2266092 0.4081954 0.3651954 0.02625552 0.02989297 0.02881464
#> 213 12.400 0.2219845 0.4081954 0.3698201 0.02612382 0.02989297 0.02896110
#> 214 12.936 0.2165702 0.4081954 0.3752344 0.02604167 0.02989297 0.02919663
#> 215 13.361 0.2067261 0.4180395 0.3752344 0.02665370 0.03089977 0.02919663
#> 216 13.936 0.0000000 0.4180395 0.5819605 0.00000000 0.03089977 0.03089977

ci <- Cuminc(time="time", status="status", data=aidssi, group="ccr5")
head(ci); tail(ci)

#>   group  time      Surv CI.1        CI.2      seSurv seCI.1      seCI.2
#> 1    WW 0.112 0.9961390    0 0.003861004 0.003853543      0 0.003853543
#> 2    WW 0.137 0.9922780    0 0.007722008 0.005439161      0 0.005439161
#> 3    WW 0.474 0.9884170    0 0.011583012 0.006648612      0 0.006648612
#> 4    WW 0.884 0.9845102    0 0.015489798 0.007684916      0 0.007684916
#> 5    WW 0.969 0.9806034    0 0.019396585 0.008590260      0 0.008590260
#> 6    WW 1.013 0.9766966    0 0.023303371 0.009402549      0 0.009402549

#>     group   time      Surv      CI.1      CI.2     seSurv     seCI.1     seCI.2
#> 210    WM 10.530 0.4846807 0.1739880 0.3413313 0.06604015 0.05007872 0.06224246
#> 211    WM 10.617 0.4667296 0.1739880 0.3592824 0.06598889 0.05007872 0.06307364
#> 212    WM 10.903 0.4487784 0.1919392 0.3592824 0.06584725 0.05208878 0.06307364
#> 213    WM 11.696 0.4283794 0.1919392 0.3796814 0.06593826 0.05208878 0.06426400
#> 214    WM 12.400 0.4079804 0.1919392 0.4000804 0.06587820 0.05208878 0.06528037
#> 215    WM 13.361 0.3626492 0.2372703 0.4000804 0.07249604 0.06544637 0.06528037

### Some fake data
fake <- data.frame(surv=c(seq(2,10,by=2),seq(1,13,by=3),seq(1,9,by=2),seq(1,13,by=3)),
                    stat=rep(0:3,5),Tstage=c(1:4,rep(1:4,rep(4,4))))
fake$stat[fake$stat==0 & fake$Tstage==2] <- 3
fake$stat[fake$stat==3 & fake$Tstage==1] <- 2
fake

#>    surv stat Tstage
#> 1     2    0      1
#> 2     4    1      2
#> 3     6    2      3
#> 4     8    3      4
#> 5    10    0      1
#> 6     1    1      1
#> 7     4    2      1
#> 8     7    2      1
#> 9    10    3      2
#> 10   13    1      2
#> 11    1    2      2
#> 12    3    3      2
#> 13    5    0      3
#> 14    7    1      3
#> 15    9    2      3
#> 16    1    3      3
#> 17    4    0      4
#> 18    7    1      4
#> 19   10    2      4
#> 20   13    3      4

Cuminc(time="surv", status="stat", data=fake)

#>   time      Surv      CI.1      CI.2      CI.3     seSurv     seCI.1     seCI.2
#> 1    1 0.8500000 0.0500000 0.0500000 0.0500000 0.07984360 0.04873397 0.04873397
#> 2    3 0.7968750 0.0500000 0.0500000 0.1031250 0.09082346 0.04873397 0.04873397
#> 3    4 0.6906250 0.1031250 0.1031250 0.1031250 0.10529847 0.06913864 0.06913864
#> 4    6 0.6278409 0.1031250 0.1659091 0.1031250 0.11290234 0.06913864 0.08822297
#> 5    7 0.4394886 0.2286932 0.2286932 0.1031250 0.12051501 0.10125712 0.10125712
#> 6    8 0.3767045 0.2286932 0.2286932 0.1659091 0.11852982 0.10125712 0.10125712
#> 7    9 0.3139205 0.2286932 0.2914773 0.1659091 0.11419868 0.10125712 0.11040510
#> 8   10 0.1883523 0.2286932 0.3542614 0.2286932 0.09708292 0.10125712 0.11658533
#> 9   13 0.0000000 0.3228693 0.3542614 0.3228693 0.00000000 0.12014817 0.11658533
#>       seCI.3
#> 1 0.04873397
#> 2 0.06913864
#> 3 0.06913864
#> 4 0.06913864
#> 5 0.06913864
#> 6 0.08822297
#> 7 0.08822297
#> 8 0.10125712
#> 9 0.12014817

# If we remove all entries with status=0,
# we should get binomial sample probabilities and corresponding SEs
fake0 <- fake[fake$stat!=0,]
Cuminc(time="surv", status="stat", data=fake0)

#>   time       Surv      CI.2      CI.3     seSurv     seCI.2     seCI.3
#> 1    1 0.87500000 0.0625000 0.0625000 0.08267973 0.06051536 0.06051536
#> 2    3 0.80769231 0.0625000 0.1298077 0.10003271 0.06051536 0.08578203
#> 3    4 0.74038462 0.1298077 0.1298077 0.11207620 0.08578203 0.08578203
#> 4    6 0.66634615 0.2038462 0.1298077 0.12291458 0.10564840 0.08578203
#> 5    7 0.59230769 0.2778846 0.1298077 0.12965260 0.11882548 0.08578203
#> 6    8 0.49358974 0.2778846 0.2285256 0.14069292 0.11882548 0.11946428
#> 7    9 0.39487179 0.3766026 0.2285256 0.14305476 0.13703651 0.11946428
#> 8   10 0.19743590 0.4753205 0.3272436 0.12190734 0.14534092 0.13736934
#> 9   13 0.09871795 0.4753205 0.4259615 0.09267128 0.14534092 0.14544561