Cut a multi-state data set at a landmark time point

Given a dataset in long format, for instance generated by msprep, this function cuts a multi-state data frame (object of type "msdata") at a landmark time point LM. Administrative censoring can be applied at time cens, equal for all individuals.

Usage

cutLMms(msdata, LM, cens)

Arguments

msdata

An object of class "msdata", such as output by msprep

LM

The landmark time point at which the cut is to be made

cens

The time point at which administrative censoring is to be applied; if missing, no administrative censoring will be applied

Value

An object of class "msdata" again, containing only follow-up data after LM. The data frame contains an extra column Tentry with the time of entry into the present state.

Details

The function has a similar purpose as the cutLM function in the dynpred package. Only follow-up after a landmark time point LM is considered, so all subjects who are no longer at risk are removed. Column time is updated based on the new Tstart and Tstop.

References

L. C. de Wreede, M. Fiocco, and H. Putter (2011). mstate: An R Package for the Analysis of Competing Risks and Multi-State Models. Journal of Statistical Software 38: 7.

Author

Hein Putter H.Putter@lumc.nl

Examples

tmat <- trans.illdeath(names=c("Tx","PR","RelDeath"))
data(ebmt3) # data from Section 4 of Putter, Fiocco & Geskus (2007)
msebmt <- msprep(time=c(NA,"prtime","rfstime"),status=c(NA,"prstat","rfsstat"),
    data=ebmt3,trans=tmat)
# Cut at 5 years
cutLMms(msebmt, LM=1826)
#> An object of class 'msdata'
#> 
#> Data:
#>        id from to trans Tstart Tstop time status Tentry
#> 62     23    1  2     1   1826  1959  133      0      0
#> 63     23    1  3     2   1826  1959  133      0      0
#> 64     24    1  2     1   1826  1837   11      0      0
#> 65     24    1  3     2   1826  1837   11      0      0
#> 71     26    2  3     3   1826  2106  280      0    136
#> 74     27    2  3     3   1826  2064  238      0    148
#> 77     28    2  3     3   1826  2042  216      0     28
#> 95     35    1  2     1   1826  2267  441      0      0
#> 96     35    1  3     2   1826  2267  441      0      0
#> 102    37    2  3     3   1826  2588  762      0    118
#> 107    39    2  3     3   1826  2182  356      0     80
#> 108    40    1  2     1   1826  1827    1      0      0
#> 109    40    1  3     2   1826  1827    1      0      0
#> 113    42    1  2     1   1826  1831    5      0      0
#> 114    42    1  3     2   1826  1831    5      0      0
#> 127    47    2  3     3   1826  2424  598      0    163
#> 133    49    2  3     3   1826  2185  359      0     41
#> 134    50    1  2     1   1826  2179  353      0      0
#> 135    50    1  3     2   1826  2179  353      0      0
#> 138    51    2  3     3   1826  1882   56      1     31
#> 141    52    2  3     3   1826  2053  227      0     41
#> 144    53    2  3     3   1826  1983  157      0     97
#> 147    54    2  3     3   1826  1969  143      0     27
#> 155    57    2  3     3   1826  1856   30      0     36
#> 166    61    2  3     3   1826  1842   16      0     37
#> 204    74    2  3     3   1826  2013  187      0     29
#> 207    75    2  3     3   1826  1841   15      0    245
#> 210    76    2  3     3   1826  1924   98      0     50
#> 235    85    1  2     1   1826  2767  941      0      0
#> 236    85    1  3     2   1826  2767  941      0      0
#> 249    90    2  3     3   1826  1838   12      0     35
#> 255    92    2  3     3   1826  1847   21      0     19
#> 261    95    1  2     1   1826  2314  488      0      0
#> 262    95    1  3     2   1826  2314  488      0      0
#> 265    97    1  2     1   1826  2260  434      0      0
#> 266    97    1  3     2   1826  2260  434      0      0
#> 267    98    1  2     1   1826  2186  360      0      0
#> 268    98    1  3     2   1826  2186  360      0      0
#> 269    99    1  2     1   1826  2232  406      0      0
#> 270    99    1  3     2   1826  2232  406      0      0
#> 278   103    1  2     1   1826  2545  719      0      0
#> 279   103    1  3     2   1826  2545  719      0      0
#> 282   105    1  2     1   1826  2447  621      0      0
#> 283   105    1  3     2   1826  2447  621      0      0
#> 312   118    2  3     3   1826  1901   75      0     48
#> 404   154    2  3     3   1826  2076  250      0    274
#> 407   155    2  3     3   1826  2163  337      1    130
#> 408   156    1  2     1   1826  2239  413      0      0
#> 409   156    1  3     2   1826  2239  413      0      0
#> 425   162    2  3     3   1826  1840   14      0    152
#> 433   165    2  3     3   1826  2058  232      0    111
#> 442   168    2  3     3   1826  2055  229      0     54
#> 467   179    2  3     3   1826  2271  445      0     80
#> 473   181    2  3     3   1826  2160  334      0     38
#> 480   184    2  3     3   1826  1936  110      0     42
#> 513   196    2  3     3   1826  2519  693      0     20
#> 516   197    2  3     3   1826  2492  666      0     19
#> 519   198    2  3     3   1826  2061  235      0     26
#> 522   199    2  3     3   1826  1896   70      1     31
#> 525   200    2  3     3   1826  2043  217      0     24
#> 534   203    2  3     3   1826  2339  513      0     21
#> 539   205    2  3     3   1826  2020  194      0     23
#> 544   207    2  3     3   1826  1831    5      0     25
#> 550   209    2  3     3   1826  1893   67      0     66
#> 587   224    1  2     1   1826  1935  109      0      0
#> 588   224    1  3     2   1826  1935  109      0      0
#> 606   231    2  3     3   1826  2652  826      0     31
#> 607   232    1  2     1   1826  1837   11      0      0
#> 608   232    1  3     2   1826  1837   11      0      0
#> 626   241    1  2     1   1826  2284  458      0      0
#> 627   241    1  3     2   1826  2284  458      1      0
#> 630   243    1  2     1   1826  2412  586      0      0
#> 631   243    1  3     2   1826  2412  586      0      0
#> 632   244    1  2     1   1826  2343  517      0      0
#> 633   244    1  3     2   1826  2343  517      0      0
#> 634   245    1  2     1   1826  2561  735      0      0
#> 635   245    1  3     2   1826  2561  735      0      0
#> 636   246    1  2     1   1826  1929  103      0      0
#> 637   246    1  3     2   1826  1929  103      1      0
#> 640   248    1  2     1   1826  2380  554      0      0
#> 641   248    1  3     2   1826  2380  554      0      0
#> 642   249    1  2     1   1826  2145  319      0      0
#> 643   249    1  3     2   1826  2145  319      0      0
#> 646   251    1  2     1   1826  2412  586      0      0
#> 647   251    1  3     2   1826  2412  586      0      0
#> 650   253    1  2     1   1826  2043  217      0      0
#> 651   253    1  3     2   1826  2043  217      0      0
#> 652   254    1  2     1   1826  2404  578      0      0
#> 653   254    1  3     2   1826  2404  578      0      0
#> 654   255    1  2     1   1826  2443  617      0      0
#> 655   255    1  3     2   1826  2443  617      0      0
#> 656   256    1  2     1   1826  2398  572      0      0
#> 657   256    1  3     2   1826  2398  572      0      0
#> 660   258    1  2     1   1826  2119  293      0      0
#> 661   258    1  3     2   1826  2119  293      0      0
#> 662   259    1  2     1   1826  2295  469      0      0
#> 663   259    1  3     2   1826  2295  469      0      0
#> 666   261    1  2     1   1826  2254  428      0      0
#> 667   261    1  3     2   1826  2254  428      0      0
#> 668   262    1  2     1   1826  2083  257      0      0
#> 669   262    1  3     2   1826  2083  257      0      0
#> 672   264    1  2     1   1826  2253  427      0      0
#> 673   264    1  3     2   1826  2253  427      0      0
#> 674   265    1  2     1   1826  2173  347      0      0
#> 675   265    1  3     2   1826  2173  347      0      0
#> 676   266    1  2     1   1826  1925   99      0      0
#> 677   266    1  3     2   1826  1925   99      0      0
#> 680   268    1  2     1   1826  2140  314      0      0
#> 681   268    1  3     2   1826  2140  314      0      0
#> 688   272    1  2     1   1826  1873   47      0      0
#> 689   272    1  3     2   1826  1873   47      0      0
#> 692   274    1  2     1   1826  1963  137      0      0
#> 693   274    1  3     2   1826  1963  137      0      0
#> 700   278    1  2     1   1826  1936  110      0      0
#> 701   278    1  3     2   1826  1936  110      0      0
#> 752   304    1  2     1   1826  1878   52      0      0
#> 753   304    1  3     2   1826  1878   52      0      0
#> 882   359    1  2     1   1826  2241  415      0      0
#> 883   359    1  3     2   1826  2241  415      0      0
#> 884   360    1  2     1   1826  2127  301      0      0
#> 885   360    1  3     2   1826  2127  301      0      0
#> 886   361    1  2     1   1826  2101  275      0      0
#> 887   361    1  3     2   1826  2101  275      0      0
#> 888   362    1  2     1   1826  2130  304      0      0
#> 889   362    1  3     2   1826  2130  304      0      0
#> 892   364    1  2     1   1826  2052  226      0      0
#> 893   364    1  3     2   1826  2052  226      0      0
#> 894   365    1  2     1   1826  1842   16      0      0
#> 895   365    1  3     2   1826  1842   16      0      0
#> 898   367    1  2     1   1826  2324  498      0      0
#> 899   367    1  3     2   1826  2324  498      1      0
#> 904   370    1  2     1   1826  2554  728      0      0
#> 905   370    1  3     2   1826  2554  728      0      0
#> 912   374    1  2     1   1826  1847   21      0      0
#> 913   374    1  3     2   1826  1847   21      0      0
#> 977   402    2  3     3   1826  2016  190      0     22
#> 982   405    1  2     1   1826  2498  672      0      0
#> 983   405    1  3     2   1826  2498  672      0      0
#> 986   407    1  2     1   1826  2463  637      0      0
#> 987   407    1  3     2   1826  2463  637      0      0
#> 988   408    1  2     1   1826  2449  623      0      0
#> 989   408    1  3     2   1826  2449  623      0      0
#> 990   409    1  2     1   1826  2437  611      0      0
#> 991   409    1  3     2   1826  2437  611      0      0
#> 992   410    1  2     1   1826  2386  560      0      0
#> 993   410    1  3     2   1826  2386  560      0      0
#> 994   411    1  2     1   1826  2204  378      0      0
#> 995   411    1  3     2   1826  2204  378      0      0
#> 996   412    1  2     1   1826  2134  308      0      0
#> 997   412    1  3     2   1826  2134  308      0      0
#> 1000  414    1  2     1   1826  1883   57      0      0
#> 1001  414    1  3     2   1826  1883   57      0      0
#> 1022  423    2  3     3   1826  2374  548      0     16
#> 1045  431    2  3     3   1826  2239  413      0     26
#> 1048  432    2  3     3   1826  2217  391      0     24
#> 1060  436    2  3     3   1826  1886   60      0     21
#> 1067  439    1  2     1   1826  1899   73      0      0
#> 1068  439    1  3     2   1826  1899   73      0      0
#> 1075  442    1  2     1   1826  1907   81      0      0
#> 1076  442    1  3     2   1826  1907   81      0      0
#> 1157  472    2  3     3   1826  2051  225      0     24
#> 1166  475    2  3     3   1826  1862   36      0     30
#> 1228  497    2  3     3   1826  2318  492      0     26
#> 1231  498    2  3     3   1826  2003  177      0     15
#> 1253  506    1  2     1   1826  2275  449      0      0
#> 1254  506    1  3     2   1826  2275  449      0      0
#> 1255  507    1  2     1   1826  2220  394      0      0
#> 1256  507    1  3     2   1826  2220  394      0      0
#> 1257  508    1  2     1   1826  2225  399      0      0
#> 1258  508    1  3     2   1826  2225  399      0      0
#> 1259  509    1  2     1   1826  1861   35      0      0
#> 1260  509    1  3     2   1826  1861   35      0      0
#> 1261  510    1  2     1   1826  1854   28      0      0
#> 1262  510    1  3     2   1826  1854   28      0      0
#> 1263  511    1  2     1   1826  1861   35      0      0
#> 1264  511    1  3     2   1826  1861   35      0      0
#> 1285  521    2  3     3   1826  2504  678      0     24
#> 1291  523    2  3     3   1826  2550  724      0     18
#> 1300  526    2  3     3   1826  2184  358      0     28
#> 1311  530    2  3     3   1826  2175  349      0    117
#> 1314  531    2  3     3   1826  2382  556      0     37
#> 1317  532    2  3     3   1826  2182  356      0     26
#> 1332  537    2  3     3   1826  2168  342      0     33
#> 1335  538    2  3     3   1826  1834    8      0     23
#> 1362  547    2  3     3   1826  2239  413      0     20
#> 1385  555    2  3     3   1826  1835    9      0     17
#> 1502  599    2  3     3   1826  2121  295      0     29
#> 1505  600    2  3     3   1826  2483  657      1     25
#> 1514  603    2  3     3   1826  1904   78      0     23
#> 1520  605    2  3     3   1826  1909   83      0     59
#> 1526  607    2  3     3   1826  1936  110      0     29
#> 1591  629    2  3     3   1826  2161  335      0     40
#> 1598  632    2  3     3   1826  1932  106      0     30
#> 1601  633    2  3     3   1826  1980  154      0     55
#> 1614  638    2  3     3   1826  1852   26      0     40
#> 1620  640    2  3     3   1826  1937  111      0     69
#> 1631  644    2  3     3   1826  1840   14      0     17
#> 1707  673    2  3     3   1826  2812  986      0     31
#> 1710  674    2  3     3   1826  2824  998      0     31
#> 1731  682    2  3     3   1826  2565  739      0     27
#> 1732  683    1  2     1   1826  2587  761      0      0
#> 1733  683    1  3     2   1826  2587  761      0      0
#> 1736  684    2  3     3   1826  2453  627      0     33
#> 1742  686    2  3     3   1826  2393  567      0     41
#> 1745  687    2  3     3   1826  2549  723      0     26
#> 1748  688    2  3     3   1826  2303  477      0    294
#> 1751  689    2  3     3   1826  2217  391      0    129
#> 1754  690    2  3     3   1826  2154  328      0     20
#> 1757  691    2  3     3   1826  2051  225      0     35
#> 1790  702    2  3     3   1826  2364  538      0     34
#> 1796  704    2  3     3   1826  2245  419      0     30
#> 1891  737    2  3     3   1826  2167  341      0     24
#> 1939  757    1  2     1   1826  1855   29      0      0
#> 1940  757    1  3     2   1826  1855   29      0      0
#> 2093  822    1  2     1   1826  1848   22      0      0
#> 2094  822    1  3     2   1826  1848   22      0      0
#> 2230  888    2  3     3   1826  2314  488      0     56
#> 2233  889    2  3     3   1826  2037  211      0     46
#> 2236  890    2  3     3   1826  2561  735      0     49
#> 2281  907    2  3     3   1826  2022  196      0     21
#> 2287  909    2  3     3   1826  1936  110      0     40
#> 2403  949    1  2     1   1826  1977  151      0      0
#> 2404  949    1  3     2   1826  1977  151      1      0
#> 2517  992    2  3     3   1826  2062  236      0     31
#> 2522  994    2  3     3   1826  1947  121      0     49
#> 2525  995    2  3     3   1826  1872   46      0     50
#> 2589 1019    2  3     3   1826  1855   29      0     25
#> 2614 1029    2  3     3   1826  1991  165      0     27
#> 2628 1034    2  3     3   1826  1831    5      0    201
#> 2658 1045    2  3     3   1826  2026  200      0     39
#> 2661 1046    2  3     3   1826  2483  657      0     18
#> 2664 1047    2  3     3   1826  2443  617      0     13
#> 2667 1048    2  3     3   1826  2001  175      0     45
#> 2670 1049    2  3     3   1826  1980  154      0     31
#> 2673 1051    1  2     1   1826  2267  441      0      0
#> 2674 1051    1  3     2   1826  2267  441      0      0
#> 2681 1055    1  2     1   1826  2038  212      0      0
#> 2682 1055    1  3     2   1826  2038  212      0      0
#> 2683 1056    1  2     1   1826  2121  295      0      0
#> 2684 1056    1  3     2   1826  2121  295      0      0
#> 2689 1059    1  2     1   1826  1875   49      0      0
#> 2690 1059    1  3     2   1826  1875   49      0      0
#> 2716 1071    2  3     3   1826  2102  276      0     17
#> 2719 1072    2  3     3   1826  2103  277      0     32
#> 2722 1073    2  3     3   1826  1942  116      0     16
#> 2769 1090    1  2     1   1826  2051  225      0      0
#> 2770 1090    1  3     2   1826  2051  225      0      0
#> 2773 1092    1  2     1   1826  2206  380      0      0
#> 2774 1092    1  3     2   1826  2206  380      0      0
#> 2787 1098    2  3     3   1826  2288  462      0     30
#> 2808 1107    2  3     3   1826  2043  217      0     97
#> 2835 1117    2  3     3   1826  2082  256      0     23
#> 2840 1119    2  3     3   1826  1923   97      0     25
#> 2843 1121    1  2     1   1826  2350  524      0      0
#> 2844 1121    1  3     2   1826  2350  524      0      0
#> 2847 1123    1  2     1   1826  2156  330      0      0
#> 2848 1123    1  3     2   1826  2156  330      0      0
#> 2853 1126    1  2     1   1826  2153  327      0      0
#> 2854 1126    1  3     2   1826  2153  327      0      0
#> 2855 1127    1  2     1   1826  1870   44      0      0
#> 2856 1127    1  3     2   1826  1870   44      0      0
#> 3038 1212    2  3     3   1826  2308  482      0     17
#> 3043 1214    2  3     3   1826  2198  372      0     15
#> 3044 1215    1  2     1   1826  2254  428      0      0
#> 3045 1215    1  3     2   1826  2254  428      0      0
#> 3046 1216    1  2     1   1826  2094  268      0      0
#> 3047 1216    1  3     2   1826  2094  268      0      0
#> 3052 1218    2  3     3   1826  1896   70      0     13
#> 3055 1219    2  3     3   1826  1882   56      0      9
#> 3108 1241    2  3     3   1826  1890   64      0     26
#> 3179 1267    2  3     3   1826  1972  146      0     21
#> 3188 1270    2  3     3   1826  1937  111      0     20
#> 3197 1273    2  3     3   1826  1970  144      0     25
#> 3200 1274    2  3     3   1826  2036  210      0     14
#> 3203 1275    2  3     3   1826  1972  146      0     21
#> 3206 1276    2  3     3   1826  1895   69      0     20
#> 3209 1277    2  3     3   1826  1841   15      0     15
#> 3215 1279    2  3     3   1826  1865   39      0     25
#> 3230 1284    2  3     3   1826  1832    6      0     15
#> 3302 1310    2  3     3   1826  2084  258      0     19
#> 3352 1329    1  2     1   1826  1968  142      0      0
#> 3353 1329    1  3     2   1826  1968  142      0      0
#> 3356 1330    2  3     3   1826  1849   23      0     33
#> 3405 1350    2  3     3   1826  2551  725      0     18
#> 3411 1352    2  3     3   1826  2492  666      0     13
#> 3414 1353    2  3     3   1826  2243  417      0     12
#> 3419 1355    2  3     3   1826  2092  266      0     15
#> 3437 1363    1  2     1   1826  2660  834      0      0
#> 3438 1363    1  3     2   1826  2660  834      0      0
#> 3441 1364    2  3     3   1826  2397  571      0     27
#> 3463 1372    2  3     3   1826  1865   39      0     51
#> 3503 1388    2  3     3   1826  2276  450      0     31
#> 3539 1403    1  2     1   1826  1980  154      0      0
#> 3540 1403    1  3     2   1826  1980  154      0      0
#> 3544 1405    1  2     1   1826  2228  402      0      0
#> 3545 1405    1  3     2   1826  2228  402      0      0
#> 3550 1408    1  2     1   1826  2052  226      0      0
#> 3551 1408    1  3     2   1826  2052  226      0      0
#> 3560 1413    1  2     1   1826  2228  402      0      0
#> 3561 1413    1  3     2   1826  2228  402      0      0
#> 3562 1414    1  2     1   1826  2016  190      0      0
#> 3563 1414    1  3     2   1826  2016  190      0      0
#> 3591 1427    2  3     3   1826  2573  747      0     29
#> 3594 1428    2  3     3   1826  2351  525      0     40
#> 3597 1429    2  3     3   1826  2466  640      0     27
#> 3600 1430    2  3     3   1826  2009  183      0     26
#> 3603 1431    2  3     3   1826  2198  372      0     22
#> 3604 1432    1  2     1   1826  1835    9      0      0
#> 3605 1432    1  3     2   1826  1835    9      0      0
#> 3612 1436    1  2     1   1826  1830    4      0      0
#> 3613 1436    1  3     2   1826  1830    4      0      0
#> 3614 1437    1  2     1   1826  1970  144      0      0
#> 3615 1437    1  3     2   1826  1970  144      0      0
#> 3616 1438    1  2     1   1826  1954  128      0      0
#> 3617 1438    1  3     2   1826  1954  128      1      0
#> 3618 1439    1  2     1   1826  1842   16      0      0
#> 3619 1439    1  3     2   1826  1842   16      0      0
#> 3620 1440    1  2     1   1826  1829    3      0      0
#> 3621 1440    1  3     2   1826  1829    3      0      0
#> 3636 1446    2  3     3   1826  2206  380      0     24
#> 3641 1448    2  3     3   1826  1838   12      0     24
#> 3646 1451    1  2     1   1826  1831    5      0      0
#> 3647 1451    1  3     2   1826  1831    5      0      0
#> 3656 1455    2  3     3   1826  2678  852      0     26
#> 3662 1457    2  3     3   1826  2265  439      0     43
#> 3666 1459    1  2     1   1826  2250  424      0      0
#> 3667 1459    1  3     2   1826  2250  424      0      0
#> 3668 1460    1  2     1   1826  2086  260      0      0
#> 3669 1460    1  3     2   1826  2086  260      0      0
#> 3676 1463    2  3     3   1826  2104  278      0     75
#> 3681 1465    2  3     3   1826  2131  305      0     26
#> 3707 1477    1  2     1   1826  2030  204      0      0
#> 3708 1477    1  3     2   1826  2030  204      0      0
#> 3714 1480    1  2     1   1826  2023  197      0      0
#> 3715 1480    1  3     2   1826  2023  197      0      0
#> 3718 1482    1  2     1   1826  2689  863      0      0
#> 3719 1482    1  3     2   1826  2689  863      0      0
#> 3726 1486    1  2     1   1826  2520  694      0      0
#> 3727 1486    1  3     2   1826  2520  694      0      0
#> 3736 1490    2  3     3   1826  2133  307      0     60
#> 3742 1492    2  3     3   1826  2224  398      0     25
#> 3745 1493    2  3     3   1826  2385  559      0     20
#> 3777 1508    2  3     3   1826  2030  204      0     58
#> 3780 1509    2  3     3   1826  2397  571      0     30
#> 3970 1582    1  2     1   1826  1926  100      0      0
#> 3971 1582    1  3     2   1826  1926  100      0      0
#> 3990 1591    1  2     1   1826  1880   54      0      0
#> 3991 1591    1  3     2   1826  1880   54      0      0
#> 4060 1616    2  3     3   1826  1964  138      0     28
#> 4067 1619    1  2     1   1826  1828    2      0      0
#> 4068 1619    1  3     2   1826  1828    2      0      0
#> 4071 1621    1  2     1   1826  2309  483      0      0
#> 4072 1621    1  3     2   1826  2309  483      0      0
#> 4079 1625    1  2     1   1826  1933  107      0      0
#> 4080 1625    1  3     2   1826  1933  107      0      0
#> 4081 1626    1  2     1   1826  2578  752      0      0
#> 4082 1626    1  3     2   1826  2578  752      0      0
#> 4085 1627    2  3     3   1826  2520  694      0     27
#> 4088 1628    2  3     3   1826  2597  771      1     22
#> 4091 1629    2  3     3   1826  2359  533      0     28
#> 4094 1630    2  3     3   1826  2380  554      0    112
#> 4104 1634    2  3     3   1826  1827    1      0     32
#> 4112 1637    2  3     3   1826  2692  866      0     29
#> 4125 1643    1  2     1   1826  2111  285      0      0
#> 4126 1643    1  3     2   1826  2111  285      1      0
#> 4166 1659    1  2     1   1826  1972  146      0      0
#> 4167 1659    1  3     2   1826  1972  146      0      0
#> 4266 1696    2  3     3   1826  2094  268      0    110
#> 4313 1714    2  3     3   1826  1837   11      0     21
#> 4358 1730    2  3     3   1826  2612  786      0     21
#> 4361 1731    2  3     3   1826  1980  154      0     20
#> 4370 1734    2  3     3   1826  1857   31      1    152
#> 4406 1747    2  3     3   1826  1890   64      0     47
#> 4409 1748    2  3     3   1826  1841   15      0     21
#> 4412 1749    2  3     3   1826  2190  364      0     42
#> 4415 1750    2  3     3   1826  2149  323      0     24
#> 4418 1751    2  3     3   1826  2128  302      0     20
#> 4426 1754    2  3     3   1826  2030  204      0     99
#> 4429 1755    2  3     3   1826  1953  127      0     25
#> 4430 1756    1  2     1   1826  1939  113      0      0
#> 4431 1756    1  3     2   1826  1939  113      0      0
#> 4476 1772    2  3     3   1826  1932  106      0     34
#> 4479 1774    1  2     1   1826  1848   22      0      0
#> 4480 1774    1  3     2   1826  1848   22      0      0
#> 4483 1775    2  3     3   1826  1933  107      0     34
#> 4486 1776    2  3     3   1826  1869   43      0     42
#> 4499 1781    2  3     3   1826  2107  281      0     44
#> 4500 1782    1  2     1   1826  2044  218      0      0
#> 4501 1782    1  3     2   1826  2044  218      0      0
#> 4504 1783    2  3     3   1826  2006  180      0     39
#> 4566 1807    2  3     3   1826  2135  309      0     29
#> 4619 1825    2  3     3   1826  1838   12      0     41
#> 4622 1826    2  3     3   1826  2046  220      0     20
#> 4632 1830    1  2     1   1826  2518  692      0      0
#> 4633 1830    1  3     2   1826  2518  692      0      0
#> 4638 1833    1  2     1   1826  2138  312      0      0
#> 4639 1833    1  3     2   1826  2138  312      0      0
#> 4640 1834    1  2     1   1826  1883   57      0      0
#> 4641 1834    1  3     2   1826  1883   57      0      0
#> 4646 1837    1  2     1   1826  1854   28      0      0
#> 4647 1837    1  3     2   1826  1854   28      0      0
#> 4656 1841    1  2     1   1826  1988  162      0      0
#> 4657 1841    1  3     2   1826  1988  162      0      0
#> 4660 1842    2  3     3   1826  2296  470      0     24
#> 4669 1845    2  3     3   1826  1828    2      0     23
#> 4689 1852    2  3     3   1826  2064  238      0     20
#> 4695 1854    2  3     3   1826  2322  496      1     26
#> 4698 1855    2  3     3   1826  2366  540      0     30
#> 4704 1857    2  3     3   1826  1867   41      1     27
#> 4707 1858    2  3     3   1826  2074  248      0     21
#> 4719 1862    2  3     3   1826  2146  320      0     26
#> 4722 1863    2  3     3   1826  2116  290      0     20
#> 4725 1864    2  3     3   1826  2182  356      0     21
#> 4728 1865    2  3     3   1826  2071  245      0     33
#> 4734 1867    2  3     3   1826  1834    8      0     24
#> 4749 1872    2  3     3   1826  1876   50      0     29
#> 4757 1875    2  3     3   1826  1832    6      0     23
#> 4799 1894    1  2     1   1826  2555  729      0      0
#> 4800 1894    1  3     2   1826  2555  729      0      0
#> 4807 1897    2  3     3   1826  2446  620      0     30
#> 4813 1899    2  3     3   1826  2044  218      0     31
#> 4816 1900    2  3     3   1826  1929  103      0     35
#> 4837 1908    2  3     3   1826  2523  697      0     21
#> 4843 1910    2  3     3   1826  2513  687      0     28
#> 4853 1914    2  3     3   1826  1963  137      0    126
#> 4971 1960    1  2     1   1826  1911   85      0      0
#> 4972 1960    1  3     2   1826  1911   85      0      0
#> 4977 1963    1  2     1   1826  2452  626      0      0
#> 4978 1963    1  3     2   1826  2452  626      0      0
#> 5011 1977    2  3     3   1826  2136  310      1     61
#> 5017 1979    2  3     3   1826  2331  505      0     31
#> 5076 2002    2  3     3   1826  2397  571      0     21
#> 5094 2010    1  2     1   1826  1839   13      0      0
#> 5095 2010    1  3     2   1826  1839   13      0      0
#> 5113 2018    1  2     1   1826  2493  667      0      0
#> 5114 2018    1  3     2   1826  2493  667      0      0
#> 5124 2023    1  2     1   1826  2252  426      0      0
#> 5125 2023    1  3     2   1826  2252  426      0      0
#> 5138 2028    2  3     3   1826  1972  146      1     14
#> 5171 2042    2  3     3   1826  2008  182      0     27
#> 5179 2045    2  3     3   1826  2393  567      0     29
#> 5210 2057    2  3     3   1826  2390  564      0     29
#> 5215 2059    2  3     3   1826  2121  295      0     27
#> 5221 2061    2  3     3   1826  2268  442      0     22
#> 5302 2091    2  3     3   1826  2270  444      0     53
#> 5323 2098    2  3     3   1826  2002  176      0     13
#> 5348 2108    1  2     1   1826  2192  366      0      0
#> 5349 2108    1  3     2   1826  2192  366      0      0
#> 5419 2137    1  2     1   1826  2208  382      0      0
#> 5420 2137    1  3     2   1826  2208  382      0      0
#> 5443 2149    1  2     1   1826  1841   15      0      0
#> 5444 2149    1  3     2   1826  1841   15      0      0
#> 5447 2151    1  2     1   1826  1828    2      0      0
#> 5448 2151    1  3     2   1826  1828    2      0      0
#> 5457 2156    1  2     1   1826  2264  438      0      0
#> 5458 2156    1  3     2   1826  2264  438      1      0
#> 5461 2157    2  3     3   1826  2345  519      0     12
#> 5467 2159    2  3     3   1826  2283  457      0     19
#> 5472 2162    1  2     1   1826  2369  543      0      0
#> 5473 2162    1  3     2   1826  2369  543      0      0
#> 5474 2163    1  2     1   1826  2352  526      0      0
#> 5475 2163    1  3     2   1826  2352  526      0      0
#> 5478 2164    2  3     3   1826  2189  363      0    165
#> 5486 2168    1  2     1   1826  1927  101      0      0
#> 5487 2168    1  3     2   1826  1927  101      0      0
#> 5488 2169    1  2     1   1826  1862   36      0      0
#> 5489 2169    1  3     2   1826  1862   36      0      0
#> 5510 2179    2  3     3   1826  2011  185      0     18
#> 5545 2192    1  2     1   1826  1952  126      0      0
#> 5546 2192    1  3     2   1826  1952  126      0      0
#> 5574 2203    2  3     3   1826  1841   15      0     41
events(cutLMms(msebmt, LM=1826))
#> $Frequencies
#>           to
#> from        Tx  PR RelDeath no event total entering
#>   Tx         0   0        7      126            133
#>   PR         0   0       10      195            205
#>   RelDeath   0   0        0       17             17
#> 
#> $Proportions
#>           to
#> from               Tx         PR   RelDeath   no event
#>   Tx       0.00000000 0.00000000 0.05263158 0.94736842
#>   PR       0.00000000 0.00000000 0.04878049 0.95121951
#>   RelDeath 0.00000000 0.00000000 0.00000000 1.00000000
#>