Make a cross-section of multi-state data at a given time point

Given a dataset in long format, for instance generated by msprep, this function takes a cross-section at a given time point, to list the subjects under observation (at risk) at that time point and the states currently occupied.

xsect(msdata, xtime = 0)

Arguments

msdata
An object of class "msdata", such as output by msprep

xtime
The time point at which the intersection is to be made

Value

A list containing idstate, a data frame containing id's and state, the number of the state currently occupied; atrisk, the number at risk, and prop, a table counting how many of those at risk occupy which state.

Details

It is possible that subjects have moved to one of the absorbing states prior to xtime; this is NOT taken into account. The function xsect only concerns subjects currently (at time) at risk.

Author

Hein Putter H.Putter@lumc.nl

Examples

tmat <- trans.illdeath(names=c("Tx","PR","RelDeath"))
data(ebmt3) # data from Section 4 of Putter, Fiocco & Geskus (2007)
msebmt <- msprep(time=c(NA,"prtime","rfstime"),status=c(NA,"prstat","rfsstat"),
    data=ebmt3,trans=tmat)
# At the start everyone is in state 1 (default xtime=0 is used)
xsect(msebmt)

#>        id state
#> 1       1     1
#> 2       2     1
#> 3       3     1
#> 4       4     1
#> 5       5     1
#> 6       6     1
#> 7       7     1
#> 8       8     1
#> 9       9     1
#> 10     10     1
#> 11     11     1
#> 12     12     1
#> 13     13     1
#> 14     14     1
#> 15     15     1
#> 16     16     1
#> 17     17     1
#> 18     18     1
#> 19     19     1
#> 20     20     1
#> 21     21     1
#> 22     22     1
#> 23     23     1
#> 24     24     1
#> 25     25     1
#> 26     26     1
#> 27     27     1
#> 28     28     1
#> 29     29     1
#> 30     30     1
#> 31     31     1
#> 32     32     1
#> 33     33     1
#> 34     34     1
#> 35     35     1
#> 36     36     1
#> 37     37     1
#> 38     38     1
#> 39     39     1
#> 40     40     1
#> 41     41     1
#> 42     42     1
#> 43     43     1
#> 44     44     1
#> 45     45     1
#> 46     46     1
#> 47     47     1
#> 48     48     1
#> 49     49     1
#> 50     50     1
#> 51     51     1
#> 52     52     1
#> 53     53     1
#> 54     54     1
#> 55     55     1
#> 56     56     1
#> 57     57     1
#> 58     58     1
#> 59     59     1
#> 60     60     1
#> 61     61     1
#> 62     62     1
#> 63     63     1
#> 64     64     1
#> 65     65     1
#> 66     66     1
#> 67     67     1
#> 68     68     1
#> 69     69     1
#> 70     70     1
#> 71     71     1
#> 72     72     1
#> 73     73     1
#> 74     74     1
#> 75     75     1
#> 76     76     1
#> 77     77     1
#> 78     78     1
#> 79     79     1
#> 80     80     1
#> 81     81     1
#> 82     82     1
#> 83     83     1
#> 84     84     1
#> 85     85     1
#> 86     86     1
#> 87     87     1
#> 88     88     1
#> 89     89     1
#> 90     90     1
#> 91     91     1
#> 92     92     1
#> 93     93     1
#> 94     94     1
#> 95     95     1
#> 96     96     1
#> 97     97     1
#> 98     98     1
#> 99     99     1
#> 100   100     1
#> 101   101     1
#> 102   102     1
#> 103   103     1
#> 104   104     1
#> 105   105     1
#> 106   106     1
#> 107   107     1
#> 108   108     1
#> 109   109     1
#> 110   110     1
#> 111   111     1
#> 112   112     1
#> 113   113     1
#> 114   114     1
#> 115   115     1
#> 116   116     1
#> 117   117     1
#> 118   118     1
#> 119   119     1
#> 120   120     1
#> 121   121     1
#> 122   122     1
#> 123   123     1
#> 124   124     1
#> 125   125     1
#> 126   126     1
#> 127   127     1
#> 128   128     1
#> 129   129     1
#> 130   130     1
#> 131   131     1
#> 132   132     1
#> 133   133     1
#> 134   134     1
#> 135   135     1
#> 136   136     1
#> 137   137     1
#> 138   138     1
#> 139   139     1
#> 140   140     1
#> 141   141     1
#> 142   142     1
#> 143   143     1
#> 144   144     1
#> 145   145     1
#> 146   146     1
#> 147   147     1
#> 148   148     1
#> 149   149     1
#> 150   150     1
#> 151   151     1
#> 152   152     1
#> 153   153     1
#> 154   154     1
#> 155   155     1
#> 156   156     1
#> 157   157     1
#> 158   158     1
#> 159   159     1
#> 160   160     1
#> 161   161     1
#> 162   162     1
#> 163   163     1
#> 164   164     1
#> 165   165     1
#> 166   166     1
#> 167   167     1
#> 168   168     1
#> 169   169     1
#> 170   170     1
#> 171   171     1
#> 172   172     1
#> 173   173     1
#> 174   174     1
#> 175   175     1
#> 176   176     1
#> 177   177     1
#> 178   178     1
#> 179   179     1
#> 180   180     1
#> 181   181     1
#> 182   182     1
#> 183   183     1
#> 184   184     1
#> 185   185     1
#> 186   186     1
#> 187   187     1
#> 188   188     1
#> 189   189     1
#> 190   190     1
#> 191   191     1
#> 192   192     1
#> 193   193     1
#> 194   194     1
#> 195   195     1
#> 196   196     1
#> 197   197     1
#> 198   198     1
#> 199   199     1
#> 200   200     1
#> 201   201     1
#> 202   202     1
#> 203   203     1
#> 204   204     1
#> 205   205     1
#> 206   206     1
#> 207   207     1
#> 208   208     1
#> 209   209     1
#> 210   210     1
#> 211   211     1
#> 212   212     1
#> 213   213     1
#> 214   214     1
#> 215   215     1
#> 216   216     1
#> 217   217     1
#> 218   218     1
#> 219   219     1
#> 220   220     1
#> 221   221     1
#> 222   222     1
#> 223   223     1
#> 224   224     1
#> 225   225     1
#> 226   226     1
#> 227   227     1
#> 228   228     1
#> 229   229     1
#> 230   230     1
#> 231   231     1
#> 232   232     1
#> 233   233     1
#> 234   234     1
#> 235   235     1
#> 236   236     1
#> 237   237     1
#> 238   238     1
#> 239   239     1
#> 240   240     1
#> 241   241     1
#> 242   242     1
#> 243   243     1
#> 244   244     1
#> 245   245     1
#> 246   246     1
#> 247   247     1
#> 248   248     1
#> 249   249     1
#> 250   250     1
#> 251   251     1
#> 252   252     1
#> 253   253     1
#> 254   254     1
#> 255   255     1
#> 256   256     1
#> 257   257     1
#> 258   258     1
#> 259   259     1
#> 260   260     1
#> 261   261     1
#> 262   262     1
#> 263   263     1
#> 264   264     1
#> 265   265     1
#> 266   266     1
#> 267   267     1
#> 268   268     1
#> 269   269     1
#> 270   270     1
#> 271   271     1
#> 272   272     1
#> 273   273     1
#> 274   274     1
#> 275   275     1
#> 276   276     1
#> 277   277     1
#> 278   278     1
#> 279   279     1
#> 280   280     1
#> 281   281     1
#> 282   282     1
#> 283   283     1
#> 284   284     1
#> 285   285     1
#> 286   286     1
#> 287   287     1
#> 288   288     1
#> 289   289     1
#> 290   290     1
#> 291   291     1
#> 292   292     1
#> 293   293     1
#> 294   294     1
#> 295   295     1
#> 296   296     1
#> 297   297     1
#> 298   298     1
#> 299   299     1
#> 300   300     1
#> 301   301     1
#> 302   302     1
#> 303   303     1
#> 304   304     1
#> 305   305     1
#> 306   306     1
#> 307   307     1
#> 308   308     1
#> 309   309     1
#> 310   310     1
#> 311   311     1
#> 312   312     1
#> 313   313     1
#> 314   314     1
#> 315   315     1
#> 316   316     1
#> 317   317     1
#> 318   318     1
#> 319   319     1
#> 320   320     1
#> 321   321     1
#> 322   322     1
#> 323   323     1
#> 324   324     1
#> 325   325     1
#> 326   326     1
#> 327   327     1
#> 328   328     1
#> 329   329     1
#> 330   330     1
#> 331   331     1
#> 332   332     1
#> 333   333     1
#> 334   334     1
#> 335   335     1
#> 336   336     1
#> 337   337     1
#> 338   338     1
#> 339   339     1
#> 340   340     1
#> 341   341     1
#> 342   342     1
#> 343   343     1
#> 344   344     1
#> 345   345     1
#> 346   346     1
#> 347   347     1
#> 348   348     1
#> 349   349     1
#> 350   350     1
#> 351   351     1
#> 352   352     1
#> 353   353     1
#> 354   354     1
#> 355   355     1
#> 356   356     1
#> 357   357     1
#> 358   358     1
#> 359   359     1
#> 360   360     1
#> 361   361     1
#> 362   362     1
#> 363   363     1
#> 364   364     1
#> 365   365     1
#> 366   366     1
#> 367   367     1
#> 368   368     1
#> 369   369     1
#> 370   370     1
#> 371   371     1
#> 372   372     1
#> 373   373     1
#> 374   374     1
#> 375   375     1
#> 376   376     1
#> 377   377     1
#> 378   378     1
#> 379   379     1
#> 380   380     1
#> 381   381     1
#> 382   382     1
#> 383   383     1
#> 384   384     1
#> 385   385     1
#> 386   386     1
#> 387   387     1
#> 388   388     1
#> 389   389     1
#> 390   390     1
#> 391   391     1
#> 392   392     1
#> 393   393     1
#> 394   394     1
#> 395   395     1
#> 396   396     1
#> 397   397     1
#> 398   398     1
#> 399   399     1
#> 400   400     1
#> 401   401     1
#> 402   402     1
#> 403   403     1
#> 404   404     1
#> 405   405     1
#> 406   406     1
#> 407   407     1
#> 408   408     1
#> 409   409     1
#> 410   410     1
#> 411   411     1
#> 412   412     1
#> 413   413     1
#> 414   414     1
#> 415   415     1
#> 416   416     1
#> 417   417     1
#> 418   418     1
#> 419   419     1
#> 420   420     1
#> 421   421     1
#> 422   422     1
#> 423   423     1
#> 424   424     1
#> 425   425     1
#> 426   426     1
#> 427   427     1
#> 428   428     1
#> 429   429     1
#> 430   430     1
#> 431   431     1
#> 432   432     1
#> 433   433     1
#> 434   434     1
#> 435   435     1
#> 436   436     1
#> 437   437     1
#> 438   438     1
#> 439   439     1
#> 440   440     1
#> 441   441     1
#> 442   442     1
#> 443   443     1
#> 444   444     1
#> 445   445     1
#> 446   446     1
#> 447   447     1
#> 448   448     1
#> 449   449     1
#> 450   450     1
#> 451   451     1
#> 452   452     1
#> 453   453     1
#> 454   454     1
#> 455   455     1
#> 456   456     1
#> 457   457     1
#> 458   458     1
#> 459   459     1
#> 460   460     1
#> 461   461     1
#> 462   462     1
#> 463   463     1
#> 464   464     1
#> 465   465     1
#> 466   466     1
#> 467   467     1
#> 468   468     1
#> 469   469     1
#> 470   470     1
#> 471   471     1
#> 472   472     1
#> 473   473     1
#> 474   474     1
#> 475   475     1
#> 476   476     1
#> 477   477     1
#> 478   478     1
#> 479   479     1
#> 480   480     1
#> 481   481     1
#> 482   482     1
#> 483   483     1
#> 484   484     1
#> 485   485     1
#> 486   486     1
#> 487   487     1
#> 488   488     1
#> 489   489     1
#> 490   490     1
#> 491   491     1
#> 492   492     1
#> 493   493     1
#> 494   494     1
#> 495   495     1
#> 496   496     1
#> 497   497     1
#> 498   498     1
#> 499   499     1
#> 500   500     1
#> 501   501     1
#> 502   502     1
#> 503   503     1
#> 504   504     1
#> 505   505     1
#> 506   506     1
#> 507   507     1
#> 508   508     1
#> 509   509     1
#> 510   510     1
#> 511   511     1
#> 512   512     1
#> 513   513     1
#> 514   514     1
#> 515   515     1
#> 516   516     1
#> 517   517     1
#> 518   518     1
#> 519   519     1
#> 520   520     1
#> 521   521     1
#> 522   522     1
#> 523   523     1
#> 524   524     1
#> 525   525     1
#> 526   526     1
#> 527   527     1
#> 528   528     1
#> 529   529     1
#> 530   530     1
#> 531   531     1
#> 532   532     1
#> 533   533     1
#> 534   534     1
#> 535   535     1
#> 536   536     1
#> 537   537     1
#> 538   538     1
#> 539   539     1
#> 540   540     1
#> 541   541     1
#> 542   542     1
#> 543   543     1
#> 544   544     1
#> 545   545     1
#> 546   546     1
#> 547   547     1
#> 548   548     1
#> 549   549     1
#> 550   550     1
#> 551   551     1
#> 552   552     1
#> 553   553     1
#> 554   554     1
#> 555   555     1
#> 556   556     1
#> 557   557     1
#> 558   558     1
#> 559   559     1
#> 560   560     1
#> 561   561     1
#> 562   562     1
#> 563   563     1
#> 564   564     1
#> 565   565     1
#> 566   566     1
#> 567   567     1
#> 568   568     1
#> 569   569     1
#> 570   570     1
#> 571   571     1
#> 572   572     1
#> 573   573     1
#> 574   574     1
#> 575   575     1
#> 576   576     1
#> 577   577     1
#> 578   578     1
#> 579   579     1
#> 580   580     1
#> 581   581     1
#> 582   582     1
#> 583   583     1
#> 584   584     1
#> 585   585     1
#> 586   586     1
#> 587   587     1
#> 588   588     1
#> 589   589     1
#> 590   590     1
#> 591   591     1
#> 592   592     1
#> 593   593     1
#> 594   594     1
#> 595   595     1
#> 596   596     1
#> 597   597     1
#> 598   598     1
#> 599   599     1
#> 600   600     1
#> 601   601     1
#> 602   602     1
#> 603   603     1
#> 604   604     1
#> 605   605     1
#> 606   606     1
#> 607   607     1
#> 608   608     1
#> 609   609     1
#> 610   610     1
#> 611   611     1
#> 612   612     1
#> 613   613     1
#> 614   614     1
#> 615   615     1
#> 616   616     1
#> 617   617     1
#> 618   618     1
#> 619   619     1
#> 620   620     1
#> 621   621     1
#> 622   622     1
#> 623   623     1
#> 624   624     1
#> 625   625     1
#> 626   626     1
#> 627   627     1
#> 628   628     1
#> 629   629     1
#> 630   630     1
#> 631   631     1
#> 632   632     1
#> 633   633     1
#> 634   634     1
#> 635   635     1
#> 636   636     1
#> 637   637     1
#> 638   638     1
#> 639   639     1
#> 640   640     1
#> 641   641     1
#> 642   642     1
#> 643   643     1
#> 644   644     1
#> 645   645     1
#> 646   646     1
#> 647   647     1
#> 648   648     1
#> 649   649     1
#> 650   650     1
#> 651   651     1
#> 652   652     1
#> 653   653     1
#> 654   654     1
#> 655   655     1
#> 656   656     1
#> 657   657     1
#> 658   658     1
#> 659   659     1
#> 660   660     1
#> 661   661     1
#> 662   662     1
#> 663   663     1
#> 664   664     1
#> 665   665     1
#> 666   666     1
#> 667   667     1
#> 668   668     1
#> 669   669     1
#> 670   670     1
#> 671   671     1
#> 672   672     1
#> 673   673     1
#> 674   674     1
#> 675   675     1
#> 676   676     1
#> 677   677     1
#> 678   678     1
#> 679   679     1
#> 680   680     1
#> 681   681     1
#> 682   682     1
#> 683   683     1
#> 684   684     1
#> 685   685     1
#> 686   686     1
#> 687   687     1
#> 688   688     1
#> 689   689     1
#> 690   690     1
#> 691   691     1
#> 692   692     1
#> 693   693     1
#> 694   694     1
#> 695   695     1
#> 696   696     1
#> 697   697     1
#> 698   698     1
#> 699   699     1
#> 700   700     1
#> 701   701     1
#> 702   702     1
#> 703   703     1
#> 704   704     1
#> 705   705     1
#> 706   706     1
#> 707   707     1
#> 708   708     1
#> 709   709     1
#> 710   710     1
#> 711   711     1
#> 712   712     1
#> 713   713     1
#> 714   714     1
#> 715   715     1
#> 716   716     1
#> 717   717     1
#> 718   718     1
#> 719   719     1
#> 720   720     1
#> 721   721     1
#> 722   722     1
#> 723   723     1
#> 724   724     1
#> 725   725     1
#> 726   726     1
#> 727   727     1
#> 728   728     1
#> 729   729     1
#> 730   730     1
#> 731   731     1
#> 732   732     1
#> 733   733     1
#> 734   734     1
#> 735   735     1
#> 736   736     1
#> 737   737     1
#> 738   738     1
#> 739   739     1
#> 740   740     1
#> 741   741     1
#> 742   742     1
#> 743   743     1
#> 744   744     1
#> 745   745     1
#> 746   746     1
#> 747   747     1
#> 748   748     1
#> 749   749     1
#> 750   750     1
#> 751   751     1
#> 752   752     1
#> 753   753     1
#> 754   754     1
#> 755   755     1
#> 756   756     1
#> 757   757     1
#> 758   758     1
#> 759   759     1
#> 760   760     1
#> 761   761     1
#> 762   762     1
#> 763   763     1
#> 764   764     1
#> 765   765     1
#> 766   766     1
#> 767   767     1
#> 768   768     1
#> 769   769     1
#> 770   770     1
#> 771   771     1
#> 772   772     1
#> 773   773     1
#> 774   774     1
#> 775   775     1
#> 776   776     1
#> 777   777     1
#> 778   778     1
#> 779   779     1
#> 780   780     1
#> 781   781     1
#> 782   782     1
#> 783   783     1
#> 784   784     1
#> 785   785     1
#> 786   786     1
#> 787   787     1
#> 788   788     1
#> 789   789     1
#> 790   790     1
#> 791   791     1
#> 792   792     1
#> 793   793     1
#> 794   794     1
#> 795   795     1
#> 796   796     1
#> 797   797     1
#> 798   798     1
#> 799   799     1
#> 800   800     1
#> 801   801     1
#> 802   802     1
#> 803   803     1
#> 804   804     1
#> 805   805     1
#> 806   806     1
#> 807   807     1
#> 808   808     1
#> 809   809     1
#> 810   810     1
#> 811   811     1
#> 812   812     1
#> 813   813     1
#> 814   814     1
#> 815   815     1
#> 816   816     1
#> 817   817     1
#> 818   818     1
#> 819   819     1
#> 820   820     1
#> 821   821     1
#> 822   822     1
#> 823   823     1
#> 824   824     1
#> 825   825     1
#> 826   826     1
#> 827   827     1
#> 828   828     1
#> 829   829     1
#> 830   830     1
#> 831   831     1
#> 832   832     1
#> 833   833     1
#> 834   834     1
#> 835   835     1
#> 836   836     1
#> 837   837     1
#> 838   838     1
#> 839   839     1
#> 840   840     1
#> 841   841     1
#> 842   842     1
#> 843   843     1
#> 844   844     1
#> 845   845     1
#> 846   846     1
#> 847   847     1
#> 848   848     1
#> 849   849     1
#> 850   850     1
#> 851   851     1
#> 852   852     1
#> 853   853     1
#> 854   854     1
#> 855   855     1
#> 856   856     1
#> 857   857     1
#> 858   858     1
#> 859   859     1
#> 860   860     1
#> 861   861     1
#> 862   862     1
#> 863   863     1
#> 864   864     1
#> 865   865     1
#> 866   866     1
#> 867   867     1
#> 868   868     1
#> 869   869     1
#> 870   870     1
#> 871   871     1
#> 872   872     1
#> 873   873     1
#> 874   874     1
#> 875   875     1
#> 876   876     1
#> 877   877     1
#> 878   878     1
#> 879   879     1
#> 880   880     1
#> 881   881     1
#> 882   882     1
#> 883   883     1
#> 884   884     1
#> 885   885     1
#> 886   886     1
#> 887   887     1
#> 888   888     1
#> 889   889     1
#> 890   890     1
#> 891   891     1
#> 892   892     1
#> 893   893     1
#> 894   894     1
#> 895   895     1
#> 896   896     1
#> 897   897     1
#> 898   898     1
#> 899   899     1
#> 900   900     1
#> 901   901     1
#> 902   902     1
#> 903   903     1
#> 904   904     1
#> 905   905     1
#> 906   906     1
#> 907   907     1
#> 908   908     1
#> 909   909     1
#> 910   910     1
#> 911   911     1
#> 912   912     1
#> 913   913     1
#> 914   914     1
#> 915   915     1
#> 916   916     1
#> 917   917     1
#> 918   918     1
#> 919   919     1
#> 920   920     1
#> 921   921     1
#> 922   922     1
#> 923   923     1
#> 924   924     1
#> 925   925     1
#> 926   926     1
#> 927   927     1
#> 928   928     1
#> 929   929     1
#> 930   930     1
#> 931   931     1
#> 932   932     1
#> 933   933     1
#> 934   934     1
#> 935   935     1
#> 936   936     1
#> 937   937     1
#> 938   938     1
#> 939   939     1
#> 940   940     1
#> 941   941     1
#> 942   942     1
#> 943   943     1
#> 944   944     1
#> 945   945     1
#> 946   946     1
#> 947   947     1
#> 948   948     1
#> 949   949     1
#> 950   950     1
#> 951   951     1
#> 952   952     1
#> 953   953     1
#> 954   954     1
#> 955   955     1
#> 956   956     1
#> 957   957     1
#> 958   958     1
#> 959   959     1
#> 960   960     1
#> 961   961     1
#> 962   962     1
#> 963   963     1
#> 964   964     1
#> 965   965     1
#> 966   966     1
#> 967   967     1
#> 968   968     1
#> 969   969     1
#> 970   970     1
#> 971   971     1
#> 972   972     1
#> 973   973     1
#> 974   974     1
#> 975   975     1
#> 976   976     1
#> 977   977     1
#> 978   978     1
#> 979   979     1
#> 980   980     1
#> 981   981     1
#> 982   982     1
#> 983   983     1
#> 984   984     1
#> 985   985     1
#> 986   986     1
#> 987   987     1
#> 988   988     1
#> 989   989     1
#> 990   990     1
#> 991   991     1
#> 992   992     1
#> 993   993     1
#> 994   994     1
#> 995   995     1
#> 996   996     1
#> 997   997     1
#> 998   998     1
#> 999   999     1
#> 1000 1000     1
#> 1001 1001     1
#> 1002 1002     1
#> 1003 1003     1
#> 1004 1004     1
#> 1005 1005     1
#> 1006 1006     1
#> 1007 1007     1
#> 1008 1008     1
#> 1009 1009     1
#> 1010 1010     1
#> 1011 1011     1
#> 1012 1012     1
#> 1013 1013     1
#> 1014 1014     1
#> 1015 1015     1
#> 1016 1016     1
#> 1017 1017     1
#> 1018 1018     1
#> 1019 1019     1
#> 1020 1020     1
#> 1021 1021     1
#> 1022 1022     1
#> 1023 1023     1
#> 1024 1024     1
#> 1025 1025     1
#> 1026 1026     1
#> 1027 1027     1
#> 1028 1028     1
#> 1029 1029     1
#> 1030 1030     1
#> 1031 1031     1
#> 1032 1032     1
#> 1033 1033     1
#> 1034 1034     1
#> 1035 1035     1
#> 1036 1036     1
#> 1037 1037     1
#> 1038 1038     1
#> 1039 1039     1
#> 1040 1040     1
#> 1041 1041     1
#> 1042 1042     1
#> 1043 1043     1
#> 1044 1044     1
#> 1045 1045     1
#> 1046 1046     1
#> 1047 1047     1
#> 1048 1048     1
#> 1049 1049     1
#> 1050 1050     1
#> 1051 1051     1
#> 1052 1052     1
#> 1053 1053     1
#> 1054 1054     1
#> 1055 1055     1
#> 1056 1056     1
#> 1057 1057     1
#> 1058 1058     1
#> 1059 1059     1
#> 1060 1060     1
#> 1061 1061     1
#> 1062 1062     1
#> 1063 1063     1
#> 1064 1064     1
#> 1065 1065     1
#> 1066 1066     1
#> 1067 1067     1
#> 1068 1068     1
#> 1069 1069     1
#> 1070 1070     1
#> 1071 1071     1
#> 1072 1072     1
#> 1073 1073     1
#> 1074 1074     1
#> 1075 1075     1
#> 1076 1076     1
#> 1077 1077     1
#> 1078 1078     1
#> 1079 1079     1
#> 1080 1080     1
#> 1081 1081     1
#> 1082 1082     1
#> 1083 1083     1
#> 1084 1084     1
#> 1085 1085     1
#> 1086 1086     1
#> 1087 1087     1
#> 1088 1088     1
#> 1089 1089     1
#> 1090 1090     1
#> 1091 1091     1
#> 1092 1092     1
#> 1093 1093     1
#> 1094 1094     1
#> 1095 1095     1
#> 1096 1096     1
#> 1097 1097     1
#> 1098 1098     1
#> 1099 1099     1
#> 1100 1100     1
#> 1101 1101     1
#> 1102 1102     1
#> 1103 1103     1
#> 1104 1104     1
#> 1105 1105     1
#> 1106 1106     1
#> 1107 1107     1
#> 1108 1108     1
#> 1109 1109     1
#> 1110 1110     1
#> 1111 1111     1
#> 1112 1112     1
#> 1113 1113     1
#> 1114 1114     1
#> 1115 1115     1
#> 1116 1116     1
#> 1117 1117     1
#> 1118 1118     1
#> 1119 1119     1
#> 1120 1120     1
#> 1121 1121     1
#> 1122 1122     1
#> 1123 1123     1
#> 1124 1124     1
#> 1125 1125     1
#> 1126 1126     1
#> 1127 1127     1
#> 1128 1128     1
#> 1129 1129     1
#> 1130 1130     1
#> 1131 1131     1
#> 1132 1132     1
#> 1133 1133     1
#> 1134 1134     1
#> 1135 1135     1
#> 1136 1136     1
#> 1137 1137     1
#> 1138 1138     1
#> 1139 1139     1
#> 1140 1140     1
#> 1141 1141     1
#> 1142 1142     1
#> 1143 1143     1
#> 1144 1144     1
#> 1145 1145     1
#> 1146 1146     1
#> 1147 1147     1
#> 1148 1148     1
#> 1149 1149     1
#> 1150 1150     1
#> 1151 1151     1
#> 1152 1152     1
#> 1153 1153     1
#> 1154 1154     1
#> 1155 1155     1
#> 1156 1156     1
#> 1157 1157     1
#> 1158 1158     1
#> 1159 1159     1
#> 1160 1160     1
#> 1161 1161     1
#> 1162 1162     1
#> 1163 1163     1
#> 1164 1164     1
#> 1165 1165     1
#> 1166 1166     1
#> 1167 1167     1
#> 1168 1168     1
#> 1169 1169     1
#> 1170 1170     1
#> 1171 1171     1
#> 1172 1172     1
#> 1173 1173     1
#> 1174 1174     1
#> 1175 1175     1
#> 1176 1176     1
#> 1177 1177     1
#> 1178 1178     1
#> 1179 1179     1
#> 1180 1180     1
#> 1181 1181     1
#> 1182 1182     1
#> 1183 1183     1
#> 1184 1184     1
#> 1185 1185     1
#> 1186 1186     1
#> 1187 1187     1
#> 1188 1188     1
#> 1189 1189     1
#> 1190 1190     1
#> 1191 1191     1
#> 1192 1192     1
#> 1193 1193     1
#> 1194 1194     1
#> 1195 1195     1
#> 1196 1196     1
#> 1197 1197     1
#> 1198 1198     1
#> 1199 1199     1
#> 1200 1200     1
#> 1201 1201     1
#> 1202 1202     1
#> 1203 1203     1
#> 1204 1204     1
#> 1205 1205     1
#> 1206 1206     1
#> 1207 1207     1
#> 1208 1208     1
#> 1209 1209     1
#> 1210 1210     1
#> 1211 1211     1
#> 1212 1212     1
#> 1213 1213     1
#> 1214 1214     1
#> 1215 1215     1
#> 1216 1216     1
#> 1217 1217     1
#> 1218 1218     1
#> 1219 1219     1
#> 1220 1220     1
#> 1221 1221     1
#> 1222 1222     1
#> 1223 1223     1
#> 1224 1224     1
#> 1225 1225     1
#> 1226 1226     1
#> 1227 1227     1
#> 1228 1228     1
#> 1229 1229     1
#> 1230 1230     1
#> 1231 1231     1
#> 1232 1232     1
#> 1233 1233     1
#> 1234 1234     1
#> 1235 1235     1
#> 1236 1236     1
#> 1237 1237     1
#> 1238 1238     1
#> 1239 1239     1
#> 1240 1240     1
#> 1241 1241     1
#> 1242 1242     1
#> 1243 1243     1
#> 1244 1244     1
#> 1245 1245     1
#> 1246 1246     1
#> 1247 1247     1
#> 1248 1248     1
#> 1249 1249     1
#> 1250 1250     1
#> 1251 1251     1
#> 1252 1252     1
#> 1253 1253     1
#> 1254 1254     1
#> 1255 1255     1
#> 1256 1256     1
#> 1257 1257     1
#> 1258 1258     1
#> 1259 1259     1
#> 1260 1260     1
#> 1261 1261     1
#> 1262 1262     1
#> 1263 1263     1
#> 1264 1264     1
#> 1265 1265     1
#> 1266 1266     1
#> 1267 1267     1
#> 1268 1268     1
#> 1269 1269     1
#> 1270 1270     1
#> 1271 1271     1
#> 1272 1272     1
#> 1273 1273     1
#> 1274 1274     1
#> 1275 1275     1
#> 1276 1276     1
#> 1277 1277     1
#> 1278 1278     1
#> 1279 1279     1
#> 1280 1280     1
#> 1281 1281     1
#> 1282 1282     1
#> 1283 1283     1
#> 1284 1284     1
#> 1285 1285     1
#> 1286 1286     1
#> 1287 1287     1
#> 1288 1288     1
#> 1289 1289     1
#> 1290 1290     1
#> 1291 1291     1
#> 1292 1292     1
#> 1293 1293     1
#> 1294 1294     1
#> 1295 1295     1
#> 1296 1296     1
#> 1297 1297     1
#> 1298 1298     1
#> 1299 1299     1
#> 1300 1300     1
#> 1301 1301     1
#> 1302 1302     1
#> 1303 1303     1
#> 1304 1304     1
#> 1305 1305     1
#> 1306 1306     1
#> 1307 1307     1
#> 1308 1308     1
#> 1309 1309     1
#> 1310 1310     1
#> 1311 1311     1
#> 1312 1312     1
#> 1313 1313     1
#> 1314 1314     1
#> 1315 1315     1
#> 1316 1316     1
#> 1317 1317     1
#> 1318 1318     1
#> 1319 1319     1
#> 1320 1320     1
#> 1321 1321     1
#> 1322 1322     1
#> 1323 1323     1
#> 1324 1324     1
#> 1325 1325     1
#> 1326 1326     1
#> 1327 1327     1
#> 1328 1328     1
#> 1329 1329     1
#> 1330 1330     1
#> 1331 1331     1
#> 1332 1332     1
#> 1333 1333     1
#> 1334 1334     1
#> 1335 1335     1
#> 1336 1336     1
#> 1337 1337     1
#> 1338 1338     1
#> 1339 1339     1
#> 1340 1340     1
#> 1341 1341     1
#> 1342 1342     1
#> 1343 1343     1
#> 1344 1344     1
#> 1345 1345     1
#> 1346 1346     1
#> 1347 1347     1
#> 1348 1348     1
#> 1349 1349     1
#> 1350 1350     1
#> 1351 1351     1
#> 1352 1352     1
#> 1353 1353     1
#> 1354 1354     1
#> 1355 1355     1
#> 1356 1356     1
#> 1357 1357     1
#> 1358 1358     1
#> 1359 1359     1
#> 1360 1360     1
#> 1361 1361     1
#> 1362 1362     1
#> 1363 1363     1
#> 1364 1364     1
#> 1365 1365     1
#> 1366 1366     1
#> 1367 1367     1
#> 1368 1368     1
#> 1369 1369     1
#> 1370 1370     1
#> 1371 1371     1
#> 1372 1372     1
#> 1373 1373     1
#> 1374 1374     1
#> 1375 1375     1
#> 1376 1376     1
#> 1377 1377     1
#> 1378 1378     1
#> 1379 1379     1
#> 1380 1380     1
#> 1381 1381     1
#> 1382 1382     1
#> 1383 1383     1
#> 1384 1384     1
#> 1385 1385     1
#> 1386 1386     1
#> 1387 1387     1
#> 1388 1388     1
#> 1389 1389     1
#> 1390 1390     1
#> 1391 1391     1
#> 1392 1392     1
#> 1393 1393     1
#> 1394 1394     1
#> 1395 1395     1
#> 1396 1396     1
#> 1397 1397     1
#> 1398 1398     1
#> 1399 1399     1
#> 1400 1400     1
#> 1401 1401     1
#> 1402 1402     1
#> 1403 1403     1
#> 1404 1404     1
#> 1405 1405     1
#> 1406 1406     1
#> 1407 1407     1
#> 1408 1408     1
#> 1409 1409     1
#> 1410 1410     1
#> 1411 1411     1
#> 1412 1412     1
#> 1413 1413     1
#> 1414 1414     1
#> 1415 1415     1
#> 1416 1416     1
#> 1417 1417     1
#> 1418 1418     1
#> 1419 1419     1
#> 1420 1420     1
#> 1421 1421     1
#> 1422 1422     1
#> 1423 1423     1
#> 1424 1424     1
#> 1425 1425     1
#> 1426 1426     1
#> 1427 1427     1
#> 1428 1428     1
#> 1429 1429     1
#> 1430 1430     1
#> 1431 1431     1
#> 1432 1432     1
#> 1433 1433     1
#> 1434 1434     1
#> 1435 1435     1
#> 1436 1436     1
#> 1437 1437     1
#> 1438 1438     1
#> 1439 1439     1
#> 1440 1440     1
#> 1441 1441     1
#> 1442 1442     1
#> 1443 1443     1
#> 1444 1444     1
#> 1445 1445     1
#> 1446 1446     1
#> 1447 1447     1
#> 1448 1448     1
#> 1449 1449     1
#> 1450 1450     1
#> 1451 1451     1
#> 1452 1452     1
#> 1453 1453     1
#> 1454 1454     1
#> 1455 1455     1
#> 1456 1456     1
#> 1457 1457     1
#> 1458 1458     1
#> 1459 1459     1
#> 1460 1460     1
#> 1461 1461     1
#> 1462 1462     1
#> 1463 1463     1
#> 1464 1464     1
#> 1465 1465     1
#> 1466 1466     1
#> 1467 1467     1
#> 1468 1468     1
#> 1469 1469     1
#> 1470 1470     1
#> 1471 1471     1
#> 1472 1472     1
#> 1473 1473     1
#> 1474 1474     1
#> 1475 1475     1
#> 1476 1476     1
#> 1477 1477     1
#> 1478 1478     1
#> 1479 1479     1
#> 1480 1480     1
#> 1481 1481     1
#> 1482 1482     1
#> 1483 1483     1
#> 1484 1484     1
#> 1485 1485     1
#> 1486 1486     1
#> 1487 1487     1
#> 1488 1488     1
#> 1489 1489     1
#> 1490 1490     1
#> 1491 1491     1
#> 1492 1492     1
#> 1493 1493     1
#> 1494 1494     1
#> 1495 1495     1
#> 1496 1496     1
#> 1497 1497     1
#> 1498 1498     1
#> 1499 1499     1
#> 1500 1500     1
#> 1501 1501     1
#> 1502 1502     1
#> 1503 1503     1
#> 1504 1504     1
#> 1505 1505     1
#> 1506 1506     1
#> 1507 1507     1
#> 1508 1508     1
#> 1509 1509     1
#> 1510 1510     1
#> 1511 1511     1
#> 1512 1512     1
#> 1513 1513     1
#> 1514 1514     1
#> 1515 1515     1
#> 1516 1516     1
#> 1517 1517     1
#> 1518 1518     1
#> 1519 1519     1
#> 1520 1520     1
#> 1521 1521     1
#> 1522 1522     1
#> 1523 1523     1
#> 1524 1524     1
#> 1525 1525     1
#> 1526 1526     1
#> 1527 1527     1
#> 1528 1528     1
#> 1529 1529     1
#> 1530 1530     1
#> 1531 1531     1
#> 1532 1532     1
#> 1533 1533     1
#> 1534 1534     1
#> 1535 1535     1
#> 1536 1536     1
#> 1537 1537     1
#> 1538 1538     1
#> 1539 1539     1
#> 1540 1540     1
#> 1541 1541     1
#> 1542 1542     1
#> 1543 1543     1
#> 1544 1544     1
#> 1545 1545     1
#> 1546 1546     1
#> 1547 1547     1
#> 1548 1548     1
#> 1549 1549     1
#> 1550 1550     1
#> 1551 1551     1
#> 1552 1552     1
#> 1553 1553     1
#> 1554 1554     1
#> 1555 1555     1
#> 1556 1556     1
#> 1557 1557     1
#> 1558 1558     1
#> 1559 1559     1
#> 1560 1560     1
#> 1561 1561     1
#> 1562 1562     1
#> 1563 1563     1
#> 1564 1564     1
#> 1565 1565     1
#> 1566 1566     1
#> 1567 1567     1
#> 1568 1568     1
#> 1569 1569     1
#> 1570 1570     1
#> 1571 1571     1
#> 1572 1572     1
#> 1573 1573     1
#> 1574 1574     1
#> 1575 1575     1
#> 1576 1576     1
#> 1577 1577     1
#> 1578 1578     1
#> 1579 1579     1
#> 1580 1580     1
#> 1581 1581     1
#> 1582 1582     1
#> 1583 1583     1
#> 1584 1584     1
#> 1585 1585     1
#> 1586 1586     1
#> 1587 1587     1
#> 1588 1588     1
#> 1589 1589     1
#> 1590 1590     1
#> 1591 1591     1
#> 1592 1592     1
#> 1593 1593     1
#> 1594 1594     1
#> 1595 1595     1
#> 1596 1596     1
#> 1597 1597     1
#> 1598 1598     1
#> 1599 1599     1
#> 1600 1600     1
#> 1601 1601     1
#> 1602 1602     1
#> 1603 1603     1
#> 1604 1604     1
#> 1605 1605     1
#> 1606 1606     1
#> 1607 1607     1
#> 1608 1608     1
#> 1609 1609     1
#> 1610 1610     1
#> 1611 1611     1
#> 1612 1612     1
#> 1613 1613     1
#> 1614 1614     1
#> 1615 1615     1
#> 1616 1616     1
#> 1617 1617     1
#> 1618 1618     1
#> 1619 1619     1
#> 1620 1620     1
#> 1621 1621     1
#> 1622 1622     1
#> 1623 1623     1
#> 1624 1624     1
#> 1625 1625     1
#> 1626 1626     1
#> 1627 1627     1
#> 1628 1628     1
#> 1629 1629     1
#> 1630 1630     1
#> 1631 1631     1
#> 1632 1632     1
#> 1633 1633     1
#> 1634 1634     1
#> 1635 1635     1
#> 1636 1636     1
#> 1637 1637     1
#> 1638 1638     1
#> 1639 1639     1
#> 1640 1640     1
#> 1641 1641     1
#> 1642 1642     1
#> 1643 1643     1
#> 1644 1644     1
#> 1645 1645     1
#> 1646 1646     1
#> 1647 1647     1
#> 1648 1648     1
#> 1649 1649     1
#> 1650 1650     1
#> 1651 1651     1
#> 1652 1652     1
#> 1653 1653     1
#> 1654 1654     1
#> 1655 1655     1
#> 1656 1656     1
#> 1657 1657     1
#> 1658 1658     1
#> 1659 1659     1
#> 1660 1660     1
#> 1661 1661     1
#> 1662 1662     1
#> 1663 1663     1
#> 1664 1664     1
#> 1665 1665     1
#> 1666 1666     1
#> 1667 1667     1
#> 1668 1668     1
#> 1669 1669     1
#> 1670 1670     1
#> 1671 1671     1
#> 1672 1672     1
#> 1673 1673     1
#> 1674 1674     1
#> 1675 1675     1
#> 1676 1676     1
#> 1677 1677     1
#> 1678 1678     1
#> 1679 1679     1
#> 1680 1680     1
#> 1681 1681     1
#> 1682 1682     1
#> 1683 1683     1
#> 1684 1684     1
#> 1685 1685     1
#> 1686 1686     1
#> 1687 1687     1
#> 1688 1688     1
#> 1689 1689     1
#> 1690 1690     1
#> 1691 1691     1
#> 1692 1692     1
#> 1693 1693     1
#> 1694 1694     1
#> 1695 1695     1
#> 1696 1696     1
#> 1697 1697     1
#> 1698 1698     1
#> 1699 1699     1
#> 1700 1700     1
#> 1701 1701     1
#> 1702 1702     1
#> 1703 1703     1
#> 1704 1704     1
#> 1705 1705     1
#> 1706 1706     1
#> 1707 1707     1
#> 1708 1708     1
#> 1709 1709     1
#> 1710 1710     1
#> 1711 1711     1
#> 1712 1712     1
#> 1713 1713     1
#> 1714 1714     1
#> 1715 1715     1
#> 1716 1716     1
#> 1717 1717     1
#> 1718 1718     1
#> 1719 1719     1
#> 1720 1720     1
#> 1721 1721     1
#> 1722 1722     1
#> 1723 1723     1
#> 1724 1724     1
#> 1725 1725     1
#> 1726 1726     1
#> 1727 1727     1
#> 1728 1728     1
#> 1729 1729     1
#> 1730 1730     1
#> 1731 1731     1
#> 1732 1732     1
#> 1733 1733     1
#> 1734 1734     1
#> 1735 1735     1
#> 1736 1736     1
#> 1737 1737     1
#> 1738 1738     1
#> 1739 1739     1
#> 1740 1740     1
#> 1741 1741     1
#> 1742 1742     1
#> 1743 1743     1
#> 1744 1744     1
#> 1745 1745     1
#> 1746 1746     1
#> 1747 1747     1
#> 1748 1748     1
#> 1749 1749     1
#> 1750 1750     1
#> 1751 1751     1
#> 1752 1752     1
#> 1753 1753     1
#> 1754 1754     1
#> 1755 1755     1
#> 1756 1756     1
#> 1757 1757     1
#> 1758 1758     1
#> 1759 1759     1
#> 1760 1760     1
#> 1761 1761     1
#> 1762 1762     1
#> 1763 1763     1
#> 1764 1764     1
#> 1765 1765     1
#> 1766 1766     1
#> 1767 1767     1
#> 1768 1768     1
#> 1769 1769     1
#> 1770 1770     1
#> 1771 1771     1
#> 1772 1772     1
#> 1773 1773     1
#> 1774 1774     1
#> 1775 1775     1
#> 1776 1776     1
#> 1777 1777     1
#> 1778 1778     1
#> 1779 1779     1
#> 1780 1780     1
#> 1781 1781     1
#> 1782 1782     1
#> 1783 1783     1
#> 1784 1784     1
#> 1785 1785     1
#> 1786 1786     1
#> 1787 1787     1
#> 1788 1788     1
#> 1789 1789     1
#> 1790 1790     1
#> 1791 1791     1
#> 1792 1792     1
#> 1793 1793     1
#> 1794 1794     1
#> 1795 1795     1
#> 1796 1796     1
#> 1797 1797     1
#> 1798 1798     1
#> 1799 1799     1
#> 1800 1800     1
#> 1801 1801     1
#> 1802 1802     1
#> 1803 1803     1
#> 1804 1804     1
#> 1805 1805     1
#> 1806 1806     1
#> 1807 1807     1
#> 1808 1808     1
#> 1809 1809     1
#> 1810 1810     1
#> 1811 1811     1
#> 1812 1812     1
#> 1813 1813     1
#> 1814 1814     1
#> 1815 1815     1
#> 1816 1816     1
#> 1817 1817     1
#> 1818 1818     1
#> 1819 1819     1
#> 1820 1820     1
#> 1821 1821     1
#> 1822 1822     1
#> 1823 1823     1
#> 1824 1824     1
#> 1825 1825     1
#> 1826 1826     1
#> 1827 1827     1
#> 1828 1828     1
#> 1829 1829     1
#> 1830 1830     1
#> 1831 1831     1
#> 1832 1832     1
#> 1833 1833     1
#> 1834 1834     1
#> 1835 1835     1
#> 1836 1836     1
#> 1837 1837     1
#> 1838 1838     1
#> 1839 1839     1
#> 1840 1840     1
#> 1841 1841     1
#> 1842 1842     1
#> 1843 1843     1
#> 1844 1844     1
#> 1845 1845     1
#> 1846 1846     1
#> 1847 1847     1
#> 1848 1848     1
#> 1849 1849     1
#> 1850 1850     1
#> 1851 1851     1
#> 1852 1852     1
#> 1853 1853     1
#> 1854 1854     1
#> 1855 1855     1
#> 1856 1856     1
#> 1857 1857     1
#> 1858 1858     1
#> 1859 1859     1
#> 1860 1860     1
#> 1861 1861     1
#> 1862 1862     1
#> 1863 1863     1
#> 1864 1864     1
#> 1865 1865     1
#> 1866 1866     1
#> 1867 1867     1
#> 1868 1868     1
#> 1869 1869     1
#> 1870 1870     1
#> 1871 1871     1
#> 1872 1872     1
#> 1873 1873     1
#> 1874 1874     1
#> 1875 1875     1
#> 1876 1876     1
#> 1877 1877     1
#> 1878 1878     1
#> 1879 1879     1
#> 1880 1880     1
#> 1881 1881     1
#> 1882 1882     1
#> 1883 1883     1
#> 1884 1884     1
#> 1885 1885     1
#> 1886 1886     1
#> 1887 1887     1
#> 1888 1888     1
#> 1889 1889     1
#> 1890 1890     1
#> 1891 1891     1
#> 1892 1892     1
#> 1893 1893     1
#> 1894 1894     1
#> 1895 1895     1
#> 1896 1896     1
#> 1897 1897     1
#> 1898 1898     1
#> 1899 1899     1
#> 1900 1900     1
#> 1901 1901     1
#> 1902 1902     1
#> 1903 1903     1
#> 1904 1904     1
#> 1905 1905     1
#> 1906 1906     1
#> 1907 1907     1
#> 1908 1908     1
#> 1909 1909     1
#> 1910 1910     1
#> 1911 1911     1
#> 1912 1912     1
#> 1913 1913     1
#> 1914 1914     1
#> 1915 1915     1
#> 1916 1916     1
#> 1917 1917     1
#> 1918 1918     1
#> 1919 1919     1
#> 1920 1920     1
#> 1921 1921     1
#> 1922 1922     1
#> 1923 1923     1
#> 1924 1924     1
#> 1925 1925     1
#> 1926 1926     1
#> 1927 1927     1
#> 1928 1928     1
#> 1929 1929     1
#> 1930 1930     1
#> 1931 1931     1
#> 1932 1932     1
#> 1933 1933     1
#> 1934 1934     1
#> 1935 1935     1
#> 1936 1936     1
#> 1937 1937     1
#> 1938 1938     1
#> 1939 1939     1
#> 1940 1940     1
#> 1941 1941     1
#> 1942 1942     1
#> 1943 1943     1
#> 1944 1944     1
#> 1945 1945     1
#> 1946 1946     1
#> 1947 1947     1
#> 1948 1948     1
#> 1949 1949     1
#> 1950 1950     1
#> 1951 1951     1
#> 1952 1952     1
#> 1953 1953     1
#> 1954 1954     1
#> 1955 1955     1
#> 1956 1956     1
#> 1957 1957     1
#> 1958 1958     1
#> 1959 1959     1
#> 1960 1960     1
#> 1961 1961     1
#> 1962 1962     1
#> 1963 1963     1
#> 1964 1964     1
#> 1965 1965     1
#> 1966 1966     1
#> 1967 1967     1
#> 1968 1968     1
#> 1969 1969     1
#> 1970 1970     1
#> 1971 1971     1
#> 1972 1972     1
#> 1973 1973     1
#> 1974 1974     1
#> 1975 1975     1
#> 1976 1976     1
#> 1977 1977     1
#> 1978 1978     1
#> 1979 1979     1
#> 1980 1980     1
#> 1981 1981     1
#> 1982 1982     1
#> 1983 1983     1
#> 1984 1984     1
#> 1985 1985     1
#> 1986 1986     1
#> 1987 1987     1
#> 1988 1988     1
#> 1989 1989     1
#> 1990 1990     1
#> 1991 1991     1
#> 1992 1992     1
#> 1993 1993     1
#> 1994 1994     1
#> 1995 1995     1
#> 1996 1996     1
#> 1997 1997     1
#> 1998 1998     1
#> 1999 1999     1
#> 2000 2000     1
#> 2001 2001     1
#> 2002 2002     1
#> 2003 2003     1
#> 2004 2004     1
#> 2005 2005     1
#> 2006 2006     1
#> 2007 2007     1
#> 2008 2008     1
#> 2009 2009     1
#> 2010 2010     1
#> 2011 2011     1
#> 2012 2012     1
#> 2013 2013     1
#> 2014 2014     1
#> 2015 2015     1
#> 2016 2016     1
#> 2017 2017     1
#> 2018 2018     1
#> 2019 2019     1
#> 2020 2020     1
#> 2021 2021     1
#> 2022 2022     1
#> 2023 2023     1
#> 2024 2024     1
#> 2025 2025     1
#> 2026 2026     1
#> 2027 2027     1
#> 2028 2028     1
#> 2029 2029     1
#> 2030 2030     1
#> 2031 2031     1
#> 2032 2032     1
#> 2033 2033     1
#> 2034 2034     1
#> 2035 2035     1
#> 2036 2036     1
#> 2037 2037     1
#> 2038 2038     1
#> 2039 2039     1
#> 2040 2040     1
#> 2041 2041     1
#> 2042 2042     1
#> 2043 2043     1
#> 2044 2044     1
#> 2045 2045     1
#> 2046 2046     1
#> 2047 2047     1
#> 2048 2048     1
#> 2049 2049     1
#> 2050 2050     1
#> 2051 2051     1
#> 2052 2052     1
#> 2053 2053     1
#> 2054 2054     1
#> 2055 2055     1
#> 2056 2056     1
#> 2057 2057     1
#> 2058 2058     1
#> 2059 2059     1
#> 2060 2060     1
#> 2061 2061     1
#> 2062 2062     1
#> 2063 2063     1
#> 2064 2064     1
#> 2065 2065     1
#> 2066 2066     1
#> 2067 2067     1
#> 2068 2068     1
#> 2069 2069     1
#> 2070 2070     1
#> 2071 2071     1
#> 2072 2072     1
#> 2073 2073     1
#> 2074 2074     1
#> 2075 2075     1
#> 2076 2076     1
#> 2077 2077     1
#> 2078 2078     1
#> 2079 2079     1
#> 2080 2080     1
#> 2081 2081     1
#> 2082 2082     1
#> 2083 2083     1
#> 2084 2084     1
#> 2085 2085     1
#> 2086 2086     1
#> 2087 2087     1
#> 2088 2088     1
#> 2089 2089     1
#> 2090 2090     1
#> 2091 2091     1
#> 2092 2092     1
#> 2093 2093     1
#> 2094 2094     1
#> 2095 2095     1
#> 2096 2096     1
#> 2097 2097     1
#> 2098 2098     1
#> 2099 2099     1
#> 2100 2100     1
#> 2101 2101     1
#> 2102 2102     1
#> 2103 2103     1
#> 2104 2104     1
#> 2105 2105     1
#> 2106 2106     1
#> 2107 2107     1
#> 2108 2108     1
#> 2109 2109     1
#> 2110 2110     1
#> 2111 2111     1
#> 2112 2112     1
#> 2113 2113     1
#> 2114 2114     1
#> 2115 2115     1
#> 2116 2116     1
#> 2117 2117     1
#> 2118 2118     1
#> 2119 2119     1
#> 2120 2120     1
#> 2121 2121     1
#> 2122 2122     1
#> 2123 2123     1
#> 2124 2124     1
#> 2125 2125     1
#> 2126 2126     1
#> 2127 2127     1
#> 2128 2128     1
#> 2129 2129     1
#> 2130 2130     1
#> 2131 2131     1
#> 2132 2132     1
#> 2133 2133     1
#> 2134 2134     1
#> 2135 2135     1
#> 2136 2136     1
#> 2137 2137     1
#> 2138 2138     1
#> 2139 2139     1
#> 2140 2140     1
#> 2141 2141     1
#> 2142 2142     1
#> 2143 2143     1
#> 2144 2144     1
#> 2145 2145     1
#> 2146 2146     1
#> 2147 2147     1
#> 2148 2148     1
#> 2149 2149     1
#> 2150 2150     1
#> 2151 2151     1
#> 2152 2152     1
#> 2153 2153     1
#> 2154 2154     1
#> 2155 2155     1
#> 2156 2156     1
#> 2157 2157     1
#> 2158 2158     1
#> 2159 2159     1
#> 2160 2160     1
#> 2161 2161     1
#> 2162 2162     1
#> 2163 2163     1
#> 2164 2164     1
#> 2165 2165     1
#> 2166 2166     1
#> 2167 2167     1
#> 2168 2168     1
#> 2169 2169     1
#> 2170 2170     1
#> 2171 2171     1
#> 2172 2172     1
#> 2173 2173     1
#> 2174 2174     1
#> 2175 2175     1
#> 2176 2176     1
#> 2177 2177     1
#> 2178 2178     1
#> 2179 2179     1
#> 2180 2180     1
#> 2181 2181     1
#> 2182 2182     1
#> 2183 2183     1
#> 2184 2184     1
#> 2185 2185     1
#> 2186 2186     1
#> 2187 2187     1
#> 2188 2188     1
#> 2189 2189     1
#> 2190 2190     1
#> 2191 2191     1
#> 2192 2192     1
#> 2193 2193     1
#> 2194 2194     1
#> 2195 2195     1
#> 2196 2196     1
#> 2197 2197     1
#> 2198 2198     1
#> 2199 2199     1
#> 2200 2200     1
#> 2201 2201     1
#> 2202 2202     1
#> 2203 2203     1
#> 2204 2204     1

# At 5 years
xsect(msebmt, xtime=1826)

#>       id state
#> 1     23     1
#> 2     24     1
#> 3     26     2
#> 4     27     2
#> 5     28     2
#> 6     35     1
#> 7     37     2
#> 8     39     2
#> 9     40     1
#> 10    42     1
#> 11    47     2
#> 12    49     2
#> 13    50     1
#> 14    51     2
#> 15    52     2
#> 16    53     2
#> 17    54     2
#> 18    57     2
#> 19    61     2
#> 20    74     2
#> 21    75     2
#> 22    76     2
#> 23    85     1
#> 24    90     2
#> 25    92     2
#> 26    95     1
#> 27    97     1
#> 28    98     1
#> 29    99     1
#> 30   103     1
#> 31   105     1
#> 32   118     2
#> 33   154     2
#> 34   155     2
#> 35   156     1
#> 36   162     2
#> 37   165     2
#> 38   168     2
#> 39   179     2
#> 40   181     2
#> 41   184     2
#> 42   196     2
#> 43   197     2
#> 44   198     2
#> 45   199     2
#> 46   200     2
#> 47   203     2
#> 48   205     2
#> 49   207     2
#> 50   209     2
#> 51   224     1
#> 52   231     2
#> 53   232     1
#> 54   241     1
#> 55   243     1
#> 56   244     1
#> 57   245     1
#> 58   246     1
#> 59   248     1
#> 60   249     1
#> 61   251     1
#> 62   253     1
#> 63   254     1
#> 64   255     1
#> 65   256     1
#> 66   258     1
#> 67   259     1
#> 68   261     1
#> 69   262     1
#> 70   264     1
#> 71   265     1
#> 72   266     1
#> 73   268     1
#> 74   272     1
#> 75   274     1
#> 76   278     1
#> 77   304     1
#> 78   359     1
#> 79   360     1
#> 80   361     1
#> 81   362     1
#> 82   364     1
#> 83   365     1
#> 84   367     1
#> 85   370     1
#> 86   374     1
#> 87   402     2
#> 88   405     1
#> 89   407     1
#> 90   408     1
#> 91   409     1
#> 92   410     1
#> 93   411     1
#> 94   412     1
#> 95   414     1
#> 96   423     2
#> 97   431     2
#> 98   432     2
#> 99   436     2
#> 100  439     1
#> 101  442     1
#> 102  472     2
#> 103  475     2
#> 104  497     2
#> 105  498     2
#> 106  506     1
#> 107  507     1
#> 108  508     1
#> 109  509     1
#> 110  510     1
#> 111  511     1
#> 112  521     2
#> 113  523     2
#> 114  526     2
#> 115  530     2
#> 116  531     2
#> 117  532     2
#> 118  537     2
#> 119  538     2
#> 120  547     2
#> 121  555     2
#> 122  599     2
#> 123  600     2
#> 124  603     2
#> 125  605     2
#> 126  607     2
#> 127  629     2
#> 128  632     2
#> 129  633     2
#> 130  638     2
#> 131  640     2
#> 132  644     2
#> 133  673     2
#> 134  674     2
#> 135  682     2
#> 136  683     1
#> 137  684     2
#> 138  686     2
#> 139  687     2
#> 140  688     2
#> 141  689     2
#> 142  690     2
#> 143  691     2
#> 144  702     2
#> 145  704     2
#> 146  737     2
#> 147  757     1
#> 148  822     1
#> 149  888     2
#> 150  889     2
#> 151  890     2
#> 152  907     2
#> 153  909     2
#> 154  949     1
#> 155  992     2
#> 156  994     2
#> 157  995     2
#> 158 1019     2
#> 159 1029     2
#> 160 1034     2
#> 161 1045     2
#> 162 1046     2
#> 163 1047     2
#> 164 1048     2
#> 165 1049     2
#> 166 1051     1
#> 167 1055     1
#> 168 1056     1
#> 169 1059     1
#> 170 1071     2
#> 171 1072     2
#> 172 1073     2
#> 173 1090     1
#> 174 1092     1
#> 175 1098     2
#> 176 1107     2
#> 177 1117     2
#> 178 1119     2
#> 179 1121     1
#> 180 1123     1
#> 181 1126     1
#> 182 1127     1
#> 183 1212     2
#> 184 1214     2
#> 185 1215     1
#> 186 1216     1
#> 187 1218     2
#> 188 1219     2
#> 189 1241     2
#> 190 1267     2
#> 191 1270     2
#> 192 1273     2
#> 193 1274     2
#> 194 1275     2
#> 195 1276     2
#> 196 1277     2
#> 197 1279     2
#> 198 1284     2
#> 199 1310     2
#> 200 1329     1
#> 201 1330     2
#> 202 1350     2
#> 203 1352     2
#> 204 1353     2
#> 205 1355     2
#> 206 1363     1
#> 207 1364     2
#> 208 1372     2
#> 209 1388     2
#> 210 1403     1
#> 211 1405     1
#> 212 1408     1
#> 213 1413     1
#> 214 1414     1
#> 215 1427     2
#> 216 1428     2
#> 217 1429     2
#> 218 1430     2
#> 219 1431     2
#> 220 1432     1
#> 221 1436     1
#> 222 1437     1
#> 223 1438     1
#> 224 1439     1
#> 225 1440     1
#> 226 1446     2
#> 227 1448     2
#> 228 1451     1
#> 229 1455     2
#> 230 1457     2
#> 231 1459     1
#> 232 1460     1
#> 233 1463     2
#> 234 1465     2
#> 235 1477     1
#> 236 1480     1
#> 237 1482     1
#> 238 1486     1
#> 239 1490     2
#> 240 1492     2
#> 241 1493     2
#> 242 1508     2
#> 243 1509     2
#> 244 1582     1
#> 245 1591     1
#> 246 1616     2
#> 247 1619     1
#> 248 1621     1
#> 249 1625     1
#> 250 1626     1
#> 251 1627     2
#> 252 1628     2
#> 253 1629     2
#> 254 1630     2
#> 255 1634     2
#> 256 1637     2
#> 257 1643     1
#> 258 1659     1
#> 259 1696     2
#> 260 1714     2
#> 261 1730     2
#> 262 1731     2
#> 263 1734     2
#> 264 1747     2
#> 265 1748     2
#> 266 1749     2
#> 267 1750     2
#> 268 1751     2
#> 269 1754     2
#> 270 1755     2
#> 271 1756     1
#> 272 1772     2
#> 273 1774     1
#> 274 1775     2
#> 275 1776     2
#> 276 1781     2
#> 277 1782     1
#> 278 1783     2
#> 279 1807     2
#> 280 1825     2
#> 281 1826     2
#> 282 1830     1
#> 283 1833     1
#> 284 1834     1
#> 285 1837     1
#> 286 1841     1
#> 287 1842     2
#> 288 1845     2
#> 289 1852     2
#> 290 1854     2
#> 291 1855     2
#> 292 1857     2
#> 293 1858     2
#> 294 1862     2
#> 295 1863     2
#> 296 1864     2
#> 297 1865     2
#> 298 1867     2
#> 299 1872     2
#> 300 1875     2
#> 301 1894     1
#> 302 1897     2
#> 303 1899     2
#> 304 1900     2
#> 305 1908     2
#> 306 1910     2
#> 307 1914     2
#> 308 1960     1
#> 309 1963     1
#> 310 1977     2
#> 311 1979     2
#> 312 2002     2
#> 313 2010     1
#> 314 2018     1
#> 315 2023     1
#> 316 2028     2
#> 317 2042     2
#> 318 2045     2
#> 319 2057     2
#> 320 2059     2
#> 321 2061     2
#> 322 2091     2
#> 323 2098     2
#> 324 2108     1
#> 325 2137     1
#> 326 2149     1
#> 327 2151     1
#> 328 2156     1
#> 329 2157     2
#> 330 2159     2
#> 331 2162     1
#> 332 2163     1
#> 333 2164     2
#> 334 2168     1
#> 335 2169     1
#> 336 2179     2
#> 337 2192     1
#> 338 2203     2